refactor(onboard): split onboard_cli.rs, remove dead interactive resolver code
Remove ~1200 lines of dead interactive resolver functions that were
only called by the removed GuidedOnboardUiRunner (resolve_provider_selection,
resolve_model_selection, resolve_api_key_env_selection, etc.).

Extract ~2100 lines of screen rendering, review display, starting point
detail, and utility functions into onboard_cli/screens.rs submodule.

Gate test-only items (SystemPromptSelection, WebSearchCredentialSelection,
apply_selected_api_key_env, etc.) with #[cfg(test)].

Result: onboard_cli.rs reduced from 5990 to 2694 lines (55% reduction).
