+++ crates/daemon/src/onboard_cli.rs
@@ -1,6 +1,3 @@
-// Many resolver functions here are temporarily dead after removing GuidedOnboardUiRunner;
-// they will be reconnected or removed in Task 8-10.
-#![allow(dead_code)]
 use std::env;
 use std::io::{self, Write};
 use std::path::{Path, PathBuf};
@@ -10,10 +7,10 @@ mod onboard_protocols;
 #[path = "onboard_workspace.rs"]
 mod onboard_workspace;
 pub mod presentation;
+#[allow(dead_code)] // Contains functions accessed by integration tests via pub re-exports
+mod screens;
+pub use screens::*;
 
-// dialoguer removed — using crossterm for tty detection only.
-#[allow(unused_imports)]
-use crossterm::tty::IsTty;
 use loongclaw_app as mvp;
 use loongclaw_contracts::SecretRef;
 use loongclaw_spec::CliResult;
@@ -22,7 +19,6 @@ use crate::onboard_finalize::{
     ConfigWritePlan, build_onboarding_success_summary_with_outcome, prepare_output_path_for_write,
     render_onboarding_success_summary_lines, resolve_backup_path, rollback_onboard_write_failure,
 };
-// Test-only onboard_finalize imports live in tests.rs directly.
 use crate::onboard_flow::{OnboardFlowController, run_guided_onboard_flow};
 pub use crate::onboard_preflight::{
     OnboardCheck, OnboardCheckLevel, OnboardNonInteractiveWarningPolicy,
@@ -37,18 +33,16 @@ use crate::onboard_preflight::{
     post_write_verification_failure_check, render_preflight_summary_screen_lines_with_progress,
     run_preflight_checks,
 };
-use crate::onboard_state::{
-    OnboardDraft, OnboardInteractionMode, OnboardOutcome, OnboardValueOrigin, OnboardWizardStep,
-};
+#[cfg(test)]
+use crate::onboard_state::OnboardInteractionMode;
+use crate::onboard_state::{OnboardDraft, OnboardOutcome, OnboardValueOrigin, OnboardWizardStep};
 pub use crate::onboard_types::OnboardingCredentialSummary;
-// Test-only onboard_web_search imports live in tests.rs directly.
+#[cfg(test)]
 use crate::onboard_web_search::{
-    configured_web_search_provider_credential_source_value,
-    configured_web_search_provider_env_name, configured_web_search_provider_secret,
-    current_web_search_provider, preferred_web_search_credential_env_default,
-    resolve_effective_web_search_default_provider, resolve_web_search_provider_recommendation,
+    current_web_search_provider, resolve_effective_web_search_default_provider,
+};
+use crate::onboard_web_search::{
     summarize_web_search_provider_credential, web_search_provider_display_name,
-    web_search_provider_has_inline_credential,
 };
 use crate::onboarding_model_policy;
 use crate::provider_credential_policy;
@@ -56,7 +50,6 @@ use mvp::tui_surface::{
     TuiCalloutTone, TuiChoiceSpec, TuiHeaderStyle, TuiScreenSpec, TuiSectionSpec,
     render_onboard_screen_spec,
 };
-// Test-only std::fs and time::OffsetDateTime imports live in tests.rs directly.
 
 pub use crate::onboard_finalize::{
     OnboardingAction, OnboardingActionKind, OnboardingDomainOutcome, OnboardingSuccessSummary,
@@ -106,16 +99,14 @@ pub enum SelectAction {
     Back,
 }
 
-// OnboardUi trait and PlainOnboardUi removed — all pre/post-flow prompts now
-// use the simple stdin/stdout helpers below.  Resolver functions keep the same
-// free-function signatures for the upcoming Task 8-10 reconnect.
-
 #[derive(Debug, Clone)]
 pub struct OnboardRuntimeContext {
     render_width: usize,
     workspace_root: Option<PathBuf>,
     codex_config_paths: Vec<PathBuf>,
+    #[allow(dead_code)]
     attended_terminal: bool,
+    #[allow(dead_code)]
     rich_prompt_ui_supported: bool,
 }
 
@@ -145,7 +136,8 @@ impl OnboardRuntimeContext {
     }
 }
 
-fn resolve_onboard_interaction_mode(
+#[cfg(test)]
+fn resolve_onboard_interaction_mode_for_test(
     non_interactive: bool,
     attended_terminal: bool,
     rich_prompt_ui_supported: bool,
@@ -157,15 +149,6 @@ fn resolve_onboard_interaction_mode(
         return OnboardInteractionMode::RichInteractive;
     }
     OnboardInteractionMode::PlainInteractive
-}
-
-#[cfg(test)]
-fn resolve_onboard_interaction_mode_for_test(
-    non_interactive: bool,
-    attended_terminal: bool,
-    rich_prompt_ui_supported: bool,
-) -> OnboardInteractionMode {
-    resolve_onboard_interaction_mode(non_interactive, attended_terminal, rich_prompt_ui_supported)
 }
 
 fn is_explicitly_accepted_non_interactive_warning(
@@ -202,8 +185,7 @@ const MEMORY_PROFILE_CHOICES: [(mvp::config::MemoryProfile, &str, &str); 3] = [
 ];
 
 // ---------------------------------------------------------------------------
-// Simple stdin/stdout prompt helpers — replaces the old OnboardPromptLineReader,
-// StdioOnboardLineReader, OnboardPromptCapture, and the OnboardUi trait impls.
+// Simple stdin/stdout prompt helpers for pre/post-flow prompts.
 // ---------------------------------------------------------------------------
 
 fn read_stdin_line() -> CliResult<String> {
@@ -215,29 +197,6 @@ fn read_stdin_line() -> CliResult<String> {
         return Err("onboarding cancelled: stdin closed".to_owned());
     }
     Ok(line)
-}
-
-fn prompt_stdin_with_default(label: &str, default: &str) -> CliResult<String> {
-    print!("{}", render_prompt_with_default_text(label, default));
-    io::stdout()
-        .flush()
-        .map_err(|e| format!("flush stdout failed: {e}"))?;
-    let line = read_stdin_line()?;
-    let trimmed = line.trim();
-    if trimmed.is_empty() {
-        Ok(default.to_owned())
-    } else {
-        Ok(trimmed.to_owned())
-    }
-}
-
-fn prompt_stdin_required(label: &str) -> CliResult<String> {
-    print!("{label}: ");
-    io::stdout()
-        .flush()
-        .map_err(|e| format!("flush stdout failed: {e}"))?;
-    let line = read_stdin_line()?;
-    Ok(line.trim().to_owned())
 }
 
 fn prompt_stdin_confirm(message: &str, default: bool) -> CliResult<bool> {
@@ -367,18 +326,6 @@ fn select_one_selected_index(
         SelectAction::Selected(index) => Ok(index),
         SelectAction::Back => Err(ONBOARD_BACK_NAVIGATION_SIGNAL.to_owned()),
     }
-}
-
-fn is_onboard_back_navigation_requested(error: &str) -> bool {
-    error == ONBOARD_BACK_NAVIGATION_SIGNAL
-}
-
-fn prompt_with_default_allowing_back(label: &str, default: &str) -> CliResult<String> {
-    let value = prompt_stdin_with_default(label, default)?;
-    if value.trim().eq_ignore_ascii_case("back") {
-        return Err(ONBOARD_BACK_NAVIGATION_SIGNAL.to_owned());
-    }
-    Ok(value)
 }
 
 fn build_onboard_entry_screen_options(options: &[OnboardEntryOption]) -> Vec<OnboardScreenOption> {
@@ -514,6 +461,7 @@ fn render_select_one_invalid_input_message(options: &[SelectOption]) -> String {
     )
 }
 
+#[cfg(test)]
 fn resolve_select_one_eof(default: Option<usize>) -> CliResult<usize> {
     default.ok_or_else(|| {
         "onboarding cancelled: stdin closed while waiting for required selection".to_owned()
@@ -534,21 +482,6 @@ fn print_stdout_message(line: impl Into<String>) -> CliResult<()> {
 
 fn is_explicit_onboard_clear_input(raw: &str) -> bool {
     raw.trim().eq_ignore_ascii_case(ONBOARD_CLEAR_INPUT_TOKEN)
-}
-
-fn prompt_optional(label: &str, current: Option<&str>) -> CliResult<Option<String>> {
-    let value = prompt_stdin_required(label)?;
-    let trimmed = value.trim();
-    if trimmed.is_empty() {
-        return Ok(current
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-            .map(str::to_owned));
-    }
-    if trimmed == "-" {
-        return Ok(None);
-    }
-    Ok(Some(trimmed.to_owned()))
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -646,6 +579,7 @@ impl GuidedPromptPath {
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[allow(dead_code)] // WebSearchProvider variant used only in test-gated render functions
 enum GuidedOnboardStep {
     Provider,
     Model,
@@ -701,7 +635,9 @@ impl ReviewFlowStyle {
     }
 }
 
+#[cfg(test)]
 #[derive(Debug, Clone, PartialEq, Eq)]
+#[allow(dead_code)]
 enum SystemPromptSelection {
     KeepCurrent,
     RestoreBuiltIn,
@@ -716,7 +652,9 @@ pub(crate) struct OnboardScreenOption {
     pub(crate) recommended: bool,
 }
 
+#[cfg(test)]
 #[derive(Debug, Clone, PartialEq, Eq)]
+#[allow(dead_code)]
 enum WebSearchCredentialSelection {
     KeepCurrent,
     ClearConfigured,
@@ -1256,11 +1194,6 @@ fn initial_draft_origin(entry_choice: OnboardEntryChoice) -> Option<OnboardValue
     }
 }
 
-// GuidedOnboardUiRunner removed — replaced by crate::onboard_tui::RatatuiOnboardRunner.
-
-// GuidedOnboardUiRunner and its impl blocks removed.
-// Interactive guided flow is now handled by crate::onboard_tui::RatatuiOnboardRunner.
-
 fn print_guided_step_boundary(step: OnboardWizardStep) -> CliResult<()> {
     println!("{}", guided_step_progress_line(step));
     Ok(())
@@ -1356,307 +1289,6 @@ pub fn build_channel_onboarding_follow_up_lines(
     lines
 }
 
-fn resolve_provider_selection(
-    options: &OnboardCommandOptions,
-    config: &mvp::config::LoongClawConfig,
-    provider_selection: &crate::migration::ProviderSelectionPlan,
-    guided_prompt_path: GuidedPromptPath,
-
-    context: &OnboardRuntimeContext,
-) -> CliResult<mvp::config::ProviderConfig> {
-    if options.non_interactive {
-        if let Some(provider_raw) = options.provider.as_deref() {
-            return resolve_provider_config_from_selector(
-                &config.provider,
-                provider_selection,
-                provider_raw,
-            );
-        }
-        if provider_selection.requires_explicit_choice {
-            let detected = provider_selection
-                .imported_choices
-                .iter()
-                .map(|choice| choice.profile_id.as_str())
-                .collect::<Vec<_>>()
-                .join(", ");
-            return Err(format!(
-                "multiple detected provider choices found ({detected}); rerun with --provider {} to choose the active provider",
-                crate::migration::provider_selection::PROVIDER_SELECTOR_PLACEHOLDER,
-            ));
-        }
-        if let Some(default_profile_id) = provider_selection.default_profile_id.as_deref() {
-            return resolve_provider_config_from_selector(
-                &config.provider,
-                provider_selection,
-                default_profile_id,
-            );
-        }
-        return Ok(crate::migration::resolve_provider_config_from_selection(
-            &config.provider,
-            provider_selection,
-            provider_selection
-                .default_kind
-                .unwrap_or(config.provider.kind),
-        ));
-    }
-
-    if !provider_selection.imported_choices.is_empty() {
-        let select_options: Vec<SelectOption> = provider_selection
-            .imported_choices
-            .iter()
-            .map(|choice| SelectOption {
-                label: provider_kind_display_name(choice.kind).to_owned(),
-                slug: choice.profile_id.clone(),
-                description: format!("source: {}, summary: {}", choice.source, choice.summary),
-                recommended: Some(choice.profile_id.as_str())
-                    == provider_selection.default_profile_id.as_deref(),
-            })
-            .collect();
-        let default_idx = if provider_selection.requires_explicit_choice {
-            None
-        } else {
-            provider_selection
-                .default_profile_id
-                .as_deref()
-                .and_then(|default_id| {
-                    provider_selection
-                        .imported_choices
-                        .iter()
-                        .position(|choice| choice.profile_id == default_id)
-                })
-        };
-        print_stdout_lines(render_provider_selection_header_lines(
-            provider_selection,
-            guided_prompt_path,
-            context.render_width,
-        ))?;
-        let idx = select_one_selected_index(
-            "Provider",
-            &select_options,
-            default_idx,
-            SelectInteractionMode::List,
-        )?;
-        let choice = provider_selection
-            .imported_choices
-            .get(idx)
-            .ok_or_else(|| format!("provider selection index {idx} out of range"))?;
-        return Ok(choice.config.clone());
-    }
-
-    // No imported choices — still use the numbered chooser so the provider
-    // step stays aligned with the rest of onboarding.
-    let default_provider_kind = options
-        .provider
-        .as_deref()
-        .and_then(parse_provider_kind)
-        .or(provider_selection.default_kind)
-        .or_else(|| {
-            provider_selection
-                .default_profile_id
-                .as_deref()
-                .and_then(parse_provider_kind)
-        })
-        .unwrap_or(config.provider.kind);
-    let provider_kinds = mvp::config::ProviderKind::all_sorted()
-        .iter()
-        .copied()
-        .filter(|kind| {
-            *kind != mvp::config::ProviderKind::Kimi
-                && *kind != mvp::config::ProviderKind::KimiCoding
-                && *kind != mvp::config::ProviderKind::Stepfun
-                && *kind != mvp::config::ProviderKind::StepPlan
-        })
-        .collect::<Vec<_>>();
-    let mut select_options: Vec<SelectOption> = provider_kinds
-        .iter()
-        .map(|kind| SelectOption {
-            label: provider_kind_display_name(*kind).to_owned(),
-            slug: provider_kind_id(*kind).to_owned(),
-            description: String::new(),
-            recommended: *kind == default_provider_kind,
-        })
-        .collect();
-    select_options.push(SelectOption {
-        label: "Kimi".to_owned(),
-        slug: "kimi".to_owned(),
-        description: "Kimi API or Kimi Coding".to_owned(),
-        recommended: default_provider_kind == mvp::config::ProviderKind::Kimi
-            || default_provider_kind == mvp::config::ProviderKind::KimiCoding,
-    });
-    select_options.push(SelectOption {
-        label: "Stepfun".to_owned(),
-        slug: "stepfun".to_owned(),
-        description: "Stepfun API or Step Plan".to_owned(),
-        recommended: default_provider_kind == mvp::config::ProviderKind::Stepfun
-            || default_provider_kind == mvp::config::ProviderKind::StepPlan,
-    });
-    select_options.sort_by(|a, b| a.label.cmp(&b.label));
-    let default_provider_slug = if matches!(
-        default_provider_kind,
-        mvp::config::ProviderKind::Kimi | mvp::config::ProviderKind::KimiCoding
-    ) {
-        "kimi"
-    } else if matches!(
-        default_provider_kind,
-        mvp::config::ProviderKind::Stepfun | mvp::config::ProviderKind::StepPlan
-    ) {
-        "stepfun"
-    } else {
-        provider_kind_id(default_provider_kind)
-    };
-    let default_idx = if provider_selection.requires_explicit_choice {
-        None
-    } else {
-        select_options
-            .iter()
-            .position(|option| option.slug == default_provider_slug)
-    };
-    print_stdout_lines(render_provider_selection_header_lines(
-        provider_selection,
-        guided_prompt_path,
-        context.render_width,
-    ))?;
-    let idx = select_one_selected_index(
-        "Provider",
-        &select_options,
-        default_idx,
-        SelectInteractionMode::List,
-    )?;
-    let selected_slug = select_options
-        .get(idx)
-        .ok_or_else(|| format!("provider selection index {idx} out of range"))?
-        .slug
-        .clone();
-
-    let kind: mvp::config::ProviderKind = if selected_slug == "kimi" {
-        let kimi_options = vec![
-            SelectOption {
-                label: "Kimi API".to_owned(),
-                slug: "kimi_api".to_owned(),
-                description: "Standard Kimi chat completion API".to_owned(),
-                recommended: true,
-            },
-            SelectOption {
-                label: "Kimi Coding".to_owned(),
-                slug: "kimi_coding".to_owned(),
-                description: "Kimi for coding tasks".to_owned(),
-                recommended: false,
-            },
-        ];
-        print_stdout_lines(vec!["Select the Kimi variant:".to_owned()])?;
-        let kimi_default_idx = Some(usize::from(
-            default_provider_kind == mvp::config::ProviderKind::KimiCoding,
-        ));
-        let sub_idx = select_one_selected_index(
-            "Kimi variant",
-            &kimi_options,
-            kimi_default_idx,
-            SelectInteractionMode::List,
-        )?;
-        let sub_slug = kimi_options
-            .get(sub_idx)
-            .ok_or_else(|| format!("kimi variant index {sub_idx} out of range"))?
-            .slug
-            .clone();
-        if sub_slug == "kimi_coding" {
-            mvp::config::ProviderKind::KimiCoding
-        } else {
-            mvp::config::ProviderKind::Kimi
-        }
-    } else if selected_slug == "stepfun" {
-        let stepfun_options = vec![
-            SelectOption {
-                label: "Stepfun API".to_owned(),
-                slug: "stepfun_api".to_owned(),
-                description: "Standard Stepfun chat completion API".to_owned(),
-                recommended: true,
-            },
-            SelectOption {
-                label: "Step Plan".to_owned(),
-                slug: "step_plan".to_owned(),
-                description: "Step Plan for specialized tasks".to_owned(),
-                recommended: false,
-            },
-        ];
-        print_stdout_lines(vec!["Select the Stepfun variant:".to_owned()])?;
-        let stepfun_default_idx = Some(usize::from(
-            default_provider_kind == mvp::config::ProviderKind::StepPlan,
-        ));
-        let sub_idx = select_one_selected_index(
-            "Stepfun variant",
-            &stepfun_options,
-            stepfun_default_idx,
-            SelectInteractionMode::List,
-        )?;
-        let sub_slug = stepfun_options
-            .get(sub_idx)
-            .ok_or_else(|| format!("stepfun variant index {sub_idx} out of range"))?
-            .slug
-            .clone();
-        if sub_slug == "step_plan" {
-            mvp::config::ProviderKind::StepPlan
-        } else {
-            mvp::config::ProviderKind::Stepfun
-        }
-    } else {
-        provider_kinds
-            .iter()
-            .find(|kind| provider_kind_id(**kind) == selected_slug)
-            .copied()
-            .ok_or_else(|| format!("provider kind not found for slug {}", selected_slug))?
-    };
-
-    let mut provider_config =
-        resolve_provider_config_from_selection(&config.provider, provider_selection, kind);
-
-    if let Some(region_info) = kind.region_endpoint_info() {
-        let configured_base_url = provider_config.base_url.as_str();
-        let default_region_idx = region_info
-            .variants
-            .iter()
-            .position(|variant| variant.base_url == configured_base_url)
-            .unwrap_or(0);
-        let region_options = region_info
-            .variants
-            .iter()
-            .enumerate()
-            .map(|(index, variant)| {
-                let is_default_variant = index == 0;
-                let label = if is_default_variant {
-                    format!("{} (default)", variant.label)
-                } else {
-                    variant.label.to_owned()
-                };
-                let slug = variant.base_url.to_owned();
-                let description = format!("endpoint: {}", variant.base_url);
-                let recommended = index == default_region_idx;
-                SelectOption {
-                    label,
-                    slug,
-                    description,
-                    recommended,
-                }
-            })
-            .collect::<Vec<_>>();
-        let region_prompt = format!("Select the {} region endpoint:", region_info.family_label);
-        print_stdout_lines(vec![region_prompt])?;
-        let region_idx = select_one_selected_index(
-            "Region",
-            &region_options,
-            Some(default_region_idx),
-            SelectInteractionMode::List,
-        )?;
-        let selected_base_url = region_options
-            .get(region_idx)
-            .ok_or_else(|| format!("region selection index {region_idx} out of range"))?
-            .slug
-            .clone();
-        provider_config.set_base_url(selected_base_url);
-    }
-
-    Ok(provider_config)
-}
-
 pub fn resolve_provider_config_from_selector(
     current_provider: &mvp::config::ProviderConfig,
     provider_selection: &crate::migration::ProviderSelectionPlan,
@@ -1747,172 +1379,7 @@ pub fn resolve_provider_config_from_selection(
     crate::migration::resolve_provider_config_from_selection(current_provider, plan, selected_kind)
 }
 
-fn resolve_model_selection(
-    options: &OnboardCommandOptions,
-    config: &mvp::config::LoongClawConfig,
-    guided_prompt_path: GuidedPromptPath,
-    available_models: &[String],
-
-    context: &OnboardRuntimeContext,
-) -> CliResult<String> {
-    let prompt_default = onboarding_model_policy::resolve_onboarding_model_prompt_default(
-        &config.provider,
-        options.model.as_deref(),
-    )?;
-
-    if options.non_interactive {
-        return Ok(prompt_default);
-    }
-
-    print_stdout_lines(render_model_selection_screen_lines_with_style(
-        config,
-        prompt_default.as_str(),
-        guided_prompt_path,
-        context.render_width,
-        true,
-        !available_models.is_empty(),
-    ))?;
-    if !available_models.is_empty() {
-        let catalog_choices = onboarding_model_policy::onboarding_model_catalog_choices(
-            prompt_default.as_str(),
-            available_models,
-        );
-        let (select_options, default_idx) = build_model_selection_options(&catalog_choices);
-        let idx = select_one_selected_index(
-            "Model",
-            &select_options,
-            default_idx,
-            SelectInteractionMode::Search,
-        )?;
-        let selected = select_options
-            .get(idx)
-            .ok_or_else(|| format!("model selection index {idx} out of range"))?;
-        if selected.slug != ONBOARD_CUSTOM_MODEL_OPTION_SLUG {
-            return Ok(selected.slug.clone());
-        }
-        let custom_model = prompt_stdin_with_default("Custom model id", prompt_default.as_str())?;
-        let trimmed = custom_model.trim();
-        if trimmed.is_empty() {
-            return Err("model cannot be empty".to_owned());
-        }
-        return Ok(trimmed.to_owned());
-    }
-    let value = prompt_stdin_with_default("Model", prompt_default.as_str())?;
-    let trimmed = value.trim();
-    if trimmed.is_empty() {
-        return Err("model cannot be empty".to_owned());
-    }
-    Ok(trimmed.to_owned())
-}
-
-async fn load_onboarding_model_catalog(
-    options: &OnboardCommandOptions,
-    config: &mvp::config::LoongClawConfig,
-) -> Vec<String> {
-    if options.non_interactive || options.skip_model_probe {
-        return Vec::new();
-    }
-    if !mvp::provider::provider_auth_ready(config).await {
-        return Vec::new();
-    }
-    mvp::provider::fetch_available_models(config)
-        .await
-        .unwrap_or_default()
-}
-
-fn build_model_selection_options(
-    catalog_choices: &onboarding_model_policy::OnboardingModelCatalogChoices,
-) -> (Vec<SelectOption>, Option<usize>) {
-    let default_idx = catalog_choices.default_index;
-    let mut options = Vec::new();
-
-    for (index, model) in catalog_choices.ordered_models.iter().enumerate() {
-        let is_default_model = default_idx == Some(index);
-        let description = if is_default_model {
-            "current or suggested default".to_owned()
-        } else {
-            String::new()
-        };
-
-        let option = SelectOption {
-            label: model.clone(),
-            slug: model.clone(),
-            description,
-            recommended: is_default_model,
-        };
-        options.push(option);
-    }
-
-    options.push(SelectOption {
-        label: "enter custom model id".to_owned(),
-        slug: ONBOARD_CUSTOM_MODEL_OPTION_SLUG.to_owned(),
-        description: "manually type any provider model id".to_owned(),
-        recommended: false,
-    });
-
-    (options, default_idx)
-}
-
-fn resolve_api_key_env_selection(
-    options: &OnboardCommandOptions,
-    config: &mvp::config::LoongClawConfig,
-    default_api_key_env: String,
-    guided_prompt_path: GuidedPromptPath,
-
-    context: &OnboardRuntimeContext,
-) -> CliResult<String> {
-    let explicit_selection = if let Some(api_key_env) = options.api_key_env.as_deref() {
-        if is_explicit_onboard_clear_input(api_key_env) {
-            return Ok(String::new());
-        }
-        let trimmed = api_key_env.trim();
-        if trimmed.is_empty() {
-            None
-        } else {
-            Some(validate_selected_provider_credential_env(config, trimmed)?)
-        }
-    } else {
-        None
-    };
-
-    if options.non_interactive {
-        return Ok(explicit_selection.unwrap_or(default_api_key_env));
-    }
-    let initial = explicit_selection
-        .as_deref()
-        .unwrap_or(default_api_key_env.as_str());
-    let example_env_name =
-        provider_credential_policy::provider_credential_env_hint(&config.provider)
-            .unwrap_or_else(|| "PROVIDER_API_KEY".to_owned());
-    loop {
-        print_stdout_lines(render_api_key_env_selection_screen_lines_with_style(
-            config,
-            default_api_key_env.as_str(),
-            initial,
-            guided_prompt_path,
-            context.render_width,
-            true,
-        ))?;
-        let value = prompt_stdin_with_default("Credential env var name", initial)?;
-        if is_explicit_onboard_clear_input(&value) {
-            return Ok(String::new());
-        }
-        let trimmed = value.trim();
-        if trimmed.is_empty() {
-            return Ok(String::new());
-        }
-        match validate_selected_provider_credential_env(config, trimmed) {
-            Ok(validated) => return Ok(validated),
-            Err(error) => {
-                print_stdout_message(error)?;
-                print_stdout_message(format!(
-                    "enter the environment variable name only, for example {example_env_name}, or type :clear to remove the env binding"
-                ))?;
-            }
-        }
-    }
-}
-
+#[cfg(test)]
 fn apply_selected_api_key_env(
     provider: &mut mvp::config::ProviderConfig,
     selected_api_key_env: String,
@@ -1961,322 +1428,7 @@ fn apply_selected_system_prompt(
     }
 }
 
-fn resolve_personality_selection(
-    options: &OnboardCommandOptions,
-    config: &mvp::config::LoongClawConfig,
-
-    context: &OnboardRuntimeContext,
-) -> CliResult<mvp::prompt::PromptPersonality> {
-    if options.non_interactive {
-        if let Some(personality_raw) = options.personality.as_deref() {
-            return parse_prompt_personality(personality_raw).ok_or_else(|| {
-                format!(
-                    "unsupported --personality value \"{personality_raw}\". supported: {}",
-                    supported_personality_list()
-                )
-            });
-        }
-        return Ok(config.cli.resolved_personality());
-    }
-
-    let default_personality = options
-        .personality
-        .as_deref()
-        .and_then(parse_prompt_personality)
-        .unwrap_or_else(|| config.cli.resolved_personality());
-
-    let personalities = [
-        (
-            mvp::prompt::PromptPersonality::CalmEngineering,
-            "calm engineering",
-            "rigorous, direct, and technically grounded",
-        ),
-        (
-            mvp::prompt::PromptPersonality::FriendlyCollab,
-            "friendly collab",
-            "warm, cooperative, and explanatory when helpful",
-        ),
-        (
-            mvp::prompt::PromptPersonality::AutonomousExecutor,
-            "autonomous executor",
-            "decisive, high-initiative, and execution-oriented",
-        ),
-    ];
-    let select_options: Vec<SelectOption> = personalities
-        .iter()
-        .map(|(p, label, desc)| SelectOption {
-            label: label.to_string(),
-            slug: prompt_personality_id(*p).to_owned(),
-            description: desc.to_string(),
-            recommended: *p == default_personality,
-        })
-        .collect();
-    let default_idx = personalities
-        .iter()
-        .position(|(p, _, _)| *p == default_personality);
-
-    print_stdout_lines(render_personality_selection_header_lines(
-        config,
-        context.render_width,
-    ))?;
-    let idx = select_one_selected_index(
-        "Personality",
-        &select_options,
-        default_idx,
-        SelectInteractionMode::List,
-    )?;
-    let (personality, _, _) = personalities
-        .get(idx)
-        .ok_or_else(|| format!("personality selection index {idx} out of range"))?;
-    Ok(*personality)
-}
-
-fn resolve_prompt_addendum_selection(
-    options: &OnboardCommandOptions,
-    config: &mvp::config::LoongClawConfig,
-
-    context: &OnboardRuntimeContext,
-) -> CliResult<Option<String>> {
-    if options.non_interactive {
-        return Ok(config.cli.system_prompt_addendum.clone());
-    }
-    print_stdout_lines(render_prompt_addendum_selection_screen_lines_with_style(
-        config,
-        context.render_width,
-        true,
-    ))?;
-    prompt_optional(
-        "Prompt addendum",
-        config.cli.system_prompt_addendum.as_deref(),
-    )
-}
-
-fn resolve_system_prompt_selection(
-    options: &OnboardCommandOptions,
-    config: &mvp::config::LoongClawConfig,
-
-    context: &OnboardRuntimeContext,
-) -> CliResult<SystemPromptSelection> {
-    if options.non_interactive {
-        if let Some(system_prompt) = options.system_prompt.as_deref() {
-            if is_explicit_onboard_clear_input(system_prompt) {
-                return Ok(SystemPromptSelection::RestoreBuiltIn);
-            }
-            let trimmed = system_prompt.trim();
-            if !trimmed.is_empty() {
-                return Ok(SystemPromptSelection::Set(trimmed.to_owned()));
-            }
-        }
-        return Ok(SystemPromptSelection::KeepCurrent);
-    }
-    let initial = options
-        .system_prompt
-        .as_deref()
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .unwrap_or(config.cli.system_prompt.as_str());
-    print_stdout_lines(render_system_prompt_selection_screen_lines_with_style(
-        config,
-        initial,
-        GuidedPromptPath::InlineOverride,
-        context.render_width,
-        true,
-    ))?;
-    let value = prompt_stdin_with_default("CLI system prompt", initial)?;
-    if is_explicit_onboard_clear_input(&value) {
-        return Ok(SystemPromptSelection::RestoreBuiltIn);
-    }
-    let trimmed = value.trim();
-    if trimmed.is_empty() || trimmed == config.cli.system_prompt.trim() {
-        return Ok(SystemPromptSelection::KeepCurrent);
-    }
-    Ok(SystemPromptSelection::Set(trimmed.to_owned()))
-}
-
-fn resolve_memory_profile_selection(
-    options: &OnboardCommandOptions,
-    config: &mvp::config::LoongClawConfig,
-    guided_prompt_path: GuidedPromptPath,
-
-    context: &OnboardRuntimeContext,
-) -> CliResult<mvp::config::MemoryProfile> {
-    if options.non_interactive {
-        if let Some(profile_raw) = options.memory_profile.as_deref() {
-            return parse_memory_profile(profile_raw).ok_or_else(|| {
-                format!(
-                    "unsupported --memory-profile value \"{profile_raw}\". supported: {}",
-                    supported_memory_profile_list()
-                )
-            });
-        }
-        return Ok(config.memory.profile);
-    }
-
-    let default_profile = options
-        .memory_profile
-        .as_deref()
-        .and_then(parse_memory_profile)
-        .unwrap_or(config.memory.profile);
-    let select_options: Vec<SelectOption> = MEMORY_PROFILE_CHOICES
-        .iter()
-        .map(|(p, label, desc)| SelectOption {
-            label: label.to_string(),
-            slug: memory_profile_id(*p).to_owned(),
-            description: desc.to_string(),
-            recommended: *p == default_profile,
-        })
-        .collect();
-    let default_idx = MEMORY_PROFILE_CHOICES
-        .iter()
-        .position(|(p, _, _)| *p == default_profile);
-
-    print_stdout_lines(render_memory_profile_selection_header_lines(
-        config,
-        guided_prompt_path,
-        context.render_width,
-    ))?;
-    let idx = select_one_selected_index(
-        "Memory profile",
-        &select_options,
-        default_idx,
-        SelectInteractionMode::List,
-    )?;
-    let (profile, _, _) = MEMORY_PROFILE_CHOICES
-        .get(idx)
-        .ok_or_else(|| format!("memory profile selection index {idx} out of range"))?;
-    Ok(*profile)
-}
-
-async fn resolve_web_search_provider_selection(
-    options: &OnboardCommandOptions,
-    config: &mvp::config::LoongClawConfig,
-    guided_prompt_path: GuidedPromptPath,
-
-    context: &OnboardRuntimeContext,
-) -> CliResult<String> {
-    let recommendation = resolve_web_search_provider_recommendation(options, config).await?;
-    let recommended_provider = recommendation.provider;
-    let default_provider =
-        resolve_effective_web_search_default_provider(options, config, &recommendation);
-
-    if options.non_interactive {
-        return Ok(default_provider.to_owned());
-    }
-
-    let screen_options = build_web_search_provider_screen_options(config, recommended_provider);
-    let select_options = select_options_from_screen_options(&screen_options);
-    let default_idx = screen_options
-        .iter()
-        .position(|option| option.key == default_provider);
-
-    print_stdout_lines(
-        render_web_search_provider_selection_screen_lines_with_style(
-            config,
-            recommended_provider,
-            default_provider,
-            recommendation.reason.as_str(),
-            guided_prompt_path,
-            context.render_width,
-            true,
-        ),
-    )?;
-    let idx = select_one_selected_index(
-        "Web search provider",
-        &select_options,
-        default_idx,
-        SelectInteractionMode::List,
-    )?;
-    let selected = select_options
-        .get(idx)
-        .ok_or_else(|| format!("web search provider selection index {idx} out of range"))?;
-    Ok(selected.slug.clone())
-}
-
-fn resolve_web_search_credential_selection(
-    options: &OnboardCommandOptions,
-    config: &mvp::config::LoongClawConfig,
-    provider: &str,
-    guided_prompt_path: GuidedPromptPath,
-    non_interactive: bool,
-
-    context: &OnboardRuntimeContext,
-) -> CliResult<WebSearchCredentialSelection> {
-    let Some(descriptor) = mvp::config::web_search_provider_descriptor(provider) else {
-        return Ok(WebSearchCredentialSelection::KeepCurrent);
-    };
-    if !descriptor.requires_api_key {
-        return Ok(WebSearchCredentialSelection::KeepCurrent);
-    }
-
-    let explicit_selection = if let Some(raw_env_name) = options.web_search_api_key_env.as_deref() {
-        if is_explicit_onboard_clear_input(raw_env_name) {
-            return Ok(WebSearchCredentialSelection::ClearConfigured);
-        }
-
-        let trimmed_env_name = raw_env_name.trim();
-        if trimmed_env_name.is_empty() {
-            None
-        } else {
-            let validated_env_name =
-                validate_selected_web_search_credential_env(provider, trimmed_env_name)?;
-            Some(validated_env_name)
-        }
-    } else {
-        None
-    };
-
-    let prompt_default = preferred_web_search_credential_env_default(config, provider);
-    if non_interactive {
-        if let Some(explicit_env_name) = explicit_selection {
-            return Ok(WebSearchCredentialSelection::UseEnv(explicit_env_name));
-        }
-
-        return Ok(if prompt_default.trim().is_empty() {
-            WebSearchCredentialSelection::KeepCurrent
-        } else {
-            WebSearchCredentialSelection::UseEnv(prompt_default)
-        });
-    }
-
-    let initial_value = explicit_selection
-        .as_deref()
-        .unwrap_or(prompt_default.as_str());
-    let example_env_name = descriptor
-        .default_api_key_env
-        .or_else(|| descriptor.api_key_env_names.first().copied())
-        .unwrap_or("WEB_SEARCH_API_KEY")
-        .to_owned();
-    loop {
-        print_stdout_lines(
-            render_web_search_credential_selection_screen_lines_with_style(
-                config,
-                provider,
-                initial_value,
-                guided_prompt_path,
-                context.render_width,
-                true,
-            ),
-        )?;
-        let value = prompt_stdin_with_default("Web search credential env var name", initial_value)?;
-        if is_explicit_onboard_clear_input(&value) {
-            return Ok(WebSearchCredentialSelection::ClearConfigured);
-        }
-        let trimmed = value.trim();
-        if trimmed.is_empty() {
-            return Ok(WebSearchCredentialSelection::KeepCurrent);
-        }
-        match validate_selected_web_search_credential_env(provider, trimmed) {
-            Ok(validated) => return Ok(WebSearchCredentialSelection::UseEnv(validated)),
-            Err(error) => {
-                print_stdout_message(error)?;
-                print_stdout_message(format!(
-                    "enter the environment variable name only, for example {example_env_name}, or type :clear to remove the configured web search credential"
-                ))?;
-            }
-        }
-    }
-}
-
+#[cfg(test)]
 fn build_web_search_provider_screen_options(
     config: &mvp::config::LoongClawConfig,
     recommended_provider: &str,
@@ -2300,6 +1452,7 @@ fn build_web_search_provider_screen_options(
         .collect()
 }
 
+#[cfg(test)]
 fn render_web_search_provider_selection_screen_lines_with_style(
     config: &mvp::config::LoongClawConfig,
     recommended_provider: &str,
@@ -2341,62 +1494,6 @@ fn render_web_search_provider_selection_screen_lines_with_style(
     )
 }
 
-fn onboard_credential_env_name_is_safe(raw: &str) -> bool {
-    let trimmed = raw.trim();
-    if trimmed.is_empty() {
-        return false;
-    }
-
-    let mut config = mvp::config::LoongClawConfig::default();
-    config.provider.api_key = Some(SecretRef::Env {
-        env: trimmed.to_owned(),
-    });
-    config.provider.api_key_env = None;
-
-    config.validate().is_ok()
-}
-
-fn normalize_onboard_credential_env_name(raw: &str) -> Option<String> {
-    let trimmed = raw.trim();
-    let is_empty = trimmed.is_empty();
-    if is_empty {
-        return None;
-    }
-
-    let is_safe = onboard_credential_env_name_is_safe(trimmed);
-    if !is_safe {
-        return None;
-    }
-
-    Some(trimmed.to_owned())
-}
-
-fn validate_selected_web_search_credential_env(
-    provider: &str,
-    selected_env_name: &str,
-) -> CliResult<String> {
-    let trimmed = selected_env_name.trim();
-    if trimmed.is_empty() {
-        return Ok(String::new());
-    }
-
-    if let Some(normalized) = normalize_onboard_credential_env_name(trimmed) {
-        return Ok(normalized);
-    }
-
-    let example_env_name = mvp::config::web_search_provider_descriptor(provider)
-        .and_then(|descriptor| {
-            descriptor
-                .default_api_key_env
-                .or_else(|| descriptor.api_key_env_names.first().copied())
-        })
-        .unwrap_or("WEB_SEARCH_API_KEY");
-
-    Err(format!(
-        "web search credential source must be an environment variable name like {example_env_name}"
-    ))
-}
-
 #[cfg(test)]
 fn apply_selected_web_search_credential(
     config: &mut mvp::config::LoongClawConfig,
@@ -2427,20 +1524,6 @@ fn apply_selected_web_search_credential(
         }
         _ => {}
     }
-}
-
-fn validate_selected_provider_credential_env(
-    config: &mvp::config::LoongClawConfig,
-    selected_env_name: &str,
-) -> CliResult<String> {
-    let trimmed = selected_env_name.trim();
-    if trimmed.is_empty() {
-        return Ok(String::new());
-    }
-
-    let mut candidate = config.clone();
-    apply_selected_api_key_env(&mut candidate.provider, trimmed.to_owned());
-    candidate.validate().map(|_| trimmed.to_owned())
 }
 
 fn non_interactive_preflight_warning_message(
@@ -3607,2384 +2690,5 @@ fn build_onboard_review_screen_spec_for_draft(
 }
 
 #[cfg(test)]
-pub(crate) fn render_onboard_wrapped_display_lines<I, S>(
-    display_lines: I,
-    width: usize,
-) -> Vec<String>
-where
-    I: IntoIterator<Item = S>,
-    S: AsRef<str>,
-{
-    display_lines
-        .into_iter()
-        .flat_map(|line| mvp::presentation::render_wrapped_display_line(line.as_ref(), width))
-        .collect()
-}
-
-#[cfg(test)]
-pub(crate) fn render_onboard_option_lines(
-    options: &[OnboardScreenOption],
-    width: usize,
-) -> Vec<String> {
-    let mut lines = Vec::new();
-    for option in options {
-        let suffix = if option.recommended {
-            " (recommended)"
-        } else {
-            ""
-        };
-        let prefix = render_onboard_option_prefix(&option.key);
-        let continuation = " ".repeat(prefix.chars().count());
-        lines.extend(
-            mvp::presentation::render_wrapped_text_line_with_continuation(
-                &prefix,
-                &continuation,
-                &format!("{}{}", option.label, suffix),
-                width,
-            ),
-        );
-        lines.extend(render_onboard_wrapped_display_lines(
-            option
-                .detail_lines
-                .iter()
-                .map(|detail| format!("    {detail}"))
-                .collect::<Vec<_>>(),
-            width,
-        ));
-    }
-    lines
-}
-
-pub(crate) fn render_default_choice_footer_line(key: &str, description: &str) -> String {
-    format!("press Enter to use default {key}, {description}")
-}
-
-fn render_prompt_with_default_text(label: &str, default: &str) -> String {
-    format!("{label} (default: {default}): ")
-}
-
-#[cfg(test)]
-fn render_onboard_option_prefix(key: &str) -> String {
-    format!("{key}) ")
-}
-
-fn render_default_input_hint_line(description: impl AsRef<str>) -> String {
-    format!("- press Enter to {}", description.as_ref())
-}
-
-fn render_clear_input_hint_line(description: impl AsRef<str>) -> String {
-    format!(
-        "- type {ONBOARD_CLEAR_INPUT_TOKEN} to {}",
-        description.as_ref()
-    )
-}
-
-fn render_model_selection_default_hint_line(
-    config: &mvp::config::LoongClawConfig,
-    prompt_default: &str,
-) -> String {
-    let prompt_default = prompt_default.trim();
-    let current_model = config.provider.model.trim();
-    if prompt_default == current_model {
-        render_default_input_hint_line("keep current model")
-    } else if prompt_default.is_empty() {
-        render_default_input_hint_line("leave the model blank")
-    } else {
-        render_default_input_hint_line(format!("use prefilled model: {prompt_default}"))
-    }
-}
-
-fn render_api_key_env_selection_default_hint_line(
-    config: &mvp::config::LoongClawConfig,
-    suggested_env: &str,
-    prompt_default: &str,
-) -> String {
-    let prompt_default =
-        provider_credential_policy::render_provider_credential_source_value(Some(prompt_default))
-            .unwrap_or_default();
-    let suggested_env =
-        provider_credential_policy::render_provider_credential_source_value(Some(suggested_env))
-            .unwrap_or_default();
-    let current_env =
-        provider_credential_policy::configured_provider_credential_env_binding(&config.provider)
-            .and_then(|binding| {
-                provider_credential_policy::render_provider_credential_source_value(Some(
-                    binding.env_name.as_str(),
-                ))
-            });
-
-    if prompt_default.is_empty() {
-        return render_default_input_hint_line("leave this blank");
-    }
-
-    if current_env
-        .as_deref()
-        .is_some_and(|current_env| current_env == prompt_default)
-    {
-        return render_default_input_hint_line("keep current source");
-    }
-
-    if !suggested_env.is_empty() && prompt_default == suggested_env {
-        return render_default_input_hint_line(format!("use suggested source: {prompt_default}"));
-    }
-
-    render_default_input_hint_line(format!("use prefilled source: {prompt_default}"))
-}
-
-fn render_web_search_credential_selection_default_hint_line(
-    config: &mvp::config::LoongClawConfig,
-    provider: &str,
-    prompt_default: &str,
-) -> String {
-    let prompt_default =
-        provider_credential_policy::render_provider_credential_source_value(Some(prompt_default))
-            .unwrap_or_default();
-    let suggested_env = mvp::config::web_search_provider_descriptor(provider)
-        .and_then(|descriptor| descriptor.default_api_key_env)
-        .and_then(|env_name| {
-            provider_credential_policy::render_provider_credential_source_value(Some(env_name))
-        })
-        .unwrap_or_default();
-    let current_env =
-        configured_web_search_provider_env_name(config, provider).and_then(|env_name| {
-            provider_credential_policy::render_provider_credential_source_value(Some(
-                env_name.as_str(),
-            ))
-        });
-
-    if prompt_default.is_empty() {
-        return render_default_input_hint_line("leave this blank");
-    }
-
-    if current_env
-        .as_deref()
-        .is_some_and(|current_env| current_env == prompt_default)
-    {
-        return render_default_input_hint_line("keep current source");
-    }
-
-    if !suggested_env.is_empty() && prompt_default == suggested_env {
-        return render_default_input_hint_line(format!("use suggested source: {prompt_default}"));
-    }
-
-    render_default_input_hint_line(format!("use prefilled source: {prompt_default}"))
-}
-
-fn render_system_prompt_selection_default_hint_line(
-    config: &mvp::config::LoongClawConfig,
-    prompt_default: &str,
-) -> String {
-    let prompt_default = prompt_default.trim();
-    let current_prompt = config.cli.system_prompt.trim();
-
-    if prompt_default == current_prompt {
-        if current_prompt.is_empty() {
-            render_default_input_hint_line("keep the built-in default")
-        } else {
-            render_default_input_hint_line("keep current prompt")
-        }
-    } else if prompt_default.is_empty() {
-        render_default_input_hint_line("keep the built-in default")
-    } else {
-        render_default_input_hint_line(format!("use prefilled prompt: {prompt_default}"))
-    }
-}
-
-fn with_default_choice_footer(
-    mut footer_lines: Vec<String>,
-    default_choice_line: Option<String>,
-) -> Vec<String> {
-    if let Some(default_choice_line) = default_choice_line {
-        footer_lines.insert(0, default_choice_line);
-    }
-    footer_lines
-}
-
-pub(crate) fn append_escape_cancel_hint(mut lines: Vec<String>) -> Vec<String> {
-    if !lines.iter().any(|line| {
-        let lower = line.to_ascii_lowercase();
-        lower.contains("esc") && lower.contains("cancel")
-    }) {
-        lines.push(ONBOARD_ESCAPE_CANCEL_HINT.to_owned());
-    }
-    lines
-}
-
-fn render_onboard_choice_screen(
-    header_style: OnboardHeaderStyle,
-    width: usize,
-    subtitle: &str,
-    title: &str,
-    step: Option<(GuidedOnboardStep, GuidedPromptPath)>,
-    intro_lines: Vec<String>,
-    options: Vec<OnboardScreenOption>,
-    footer_lines: Vec<String>,
-    show_escape_cancel_hint: bool,
-    color_enabled: bool,
-) -> Vec<String> {
-    let spec = build_onboard_choice_screen_spec(
-        header_style,
-        subtitle,
-        title,
-        step,
-        intro_lines,
-        options,
-        footer_lines,
-        show_escape_cancel_hint,
-    );
-
-    render_onboard_screen_spec(&spec, width, color_enabled)
-}
-
-fn render_onboard_input_screen(
-    width: usize,
-    title: &str,
-    step: GuidedOnboardStep,
-    guided_prompt_path: GuidedPromptPath,
-    context_lines: Vec<String>,
-    hint_lines: Vec<String>,
-    color_enabled: bool,
-) -> Vec<String> {
-    let spec =
-        build_onboard_input_screen_spec(title, step, guided_prompt_path, context_lines, hint_lines);
-
-    render_onboard_screen_spec(&spec, width, color_enabled)
-}
-
-pub fn render_continue_current_setup_screen_lines(
-    config: &mvp::config::LoongClawConfig,
-    width: usize,
-) -> Vec<String> {
-    render_onboard_shortcut_screen_lines_with_style(
-        OnboardShortcutKind::CurrentSetup,
-        config,
-        None,
-        width,
-        false,
-    )
-}
-
-pub fn render_continue_detected_setup_screen_lines(
-    config: &mvp::config::LoongClawConfig,
-    import_source: &str,
-    width: usize,
-) -> Vec<String> {
-    render_onboard_shortcut_screen_lines_with_style(
-        OnboardShortcutKind::DetectedSetup,
-        config,
-        Some(import_source),
-        width,
-        false,
-    )
-}
-
-fn render_onboard_shortcut_screen_lines_with_style(
-    shortcut_kind: OnboardShortcutKind,
-    config: &mvp::config::LoongClawConfig,
-    import_source: Option<&str>,
-    width: usize,
-    color_enabled: bool,
-) -> Vec<String> {
-    let spec = build_onboard_shortcut_screen_spec(shortcut_kind, config, import_source, true);
-    render_onboard_screen_spec(&spec, width, color_enabled)
-}
-
-fn render_onboard_shortcut_header_lines_with_style(
-    shortcut_kind: OnboardShortcutKind,
-    config: &mvp::config::LoongClawConfig,
-    import_source: Option<&str>,
-    width: usize,
-    color_enabled: bool,
-) -> Vec<String> {
-    let spec = build_onboard_shortcut_screen_spec(shortcut_kind, config, import_source, false);
-    render_onboard_screen_spec(&spec, width, color_enabled)
-}
-
-fn render_shortcut_default_choice_footer_line(shortcut_kind: OnboardShortcutKind) -> String {
-    render_default_choice_footer_line("1", shortcut_kind.default_choice_description())
-}
-
-pub fn render_onboarding_risk_screen_lines(width: usize) -> Vec<String> {
-    render_onboarding_risk_screen_lines_with_style(width, false)
-}
-
-fn render_onboarding_risk_screen_lines_with_style(
-    width: usize,
-    color_enabled: bool,
-) -> Vec<String> {
-    let copy = presentation::risk_screen_copy();
-    let footer_lines = append_escape_cancel_hint(vec![render_default_choice_footer_line(
-        "n",
-        copy.default_choice_description,
-    )]);
-    let spec = TuiScreenSpec {
-        header_style: TuiHeaderStyle::Brand,
-        subtitle: Some(copy.subtitle.to_owned()),
-        title: Some(copy.title.to_owned()),
-        progress_line: None,
-        intro_lines: vec!["review the trust boundary before writing any config".to_owned()],
-        sections: vec![
-            TuiSectionSpec::Callout {
-                tone: TuiCalloutTone::Warning,
-                title: Some("what onboarding can do".to_owned()),
-                lines: vec![
-                    "LoongClaw can invoke tools and read local files when enabled.".to_owned(),
-                    "Keep credentials in environment variables, not in prompts.".to_owned(),
-                    "Prefer allowlist-style tool policy for shared environments.".to_owned(),
-                ],
-            },
-            TuiSectionSpec::Narrative {
-                title: Some("recommended baseline".to_owned()),
-                lines: vec![
-                    "start with the narrowest tool scope that still lets you verify first success"
-                        .to_owned(),
-                    "you can widen channels, models, and local automation after doctor and review"
-                        .to_owned(),
-                ],
-            },
-        ],
-        choices: vec![
-            TuiChoiceSpec {
-                key: "y".to_owned(),
-                label: copy.continue_label.to_owned(),
-                detail_lines: vec![copy.continue_detail.to_owned()],
-                recommended: false,
-            },
-            TuiChoiceSpec {
-                key: "n".to_owned(),
-                label: copy.cancel_label.to_owned(),
-                detail_lines: vec![copy.cancel_detail.to_owned()],
-                recommended: false,
-            },
-        ],
-        footer_lines,
-    };
-
-    render_onboard_screen_spec(&spec, width, color_enabled)
-}
-
-fn build_onboard_shortcut_screen_spec(
-    shortcut_kind: OnboardShortcutKind,
-    config: &mvp::config::LoongClawConfig,
-    import_source: Option<&str>,
-    include_choices: bool,
-) -> TuiScreenSpec {
-    let mut snapshot_lines = Vec::new();
-    if let Some(source) = import_source {
-        let starting_point_label = onboard_starting_point_label(None, source);
-        snapshot_lines.push(onboard_display_line(
-            "- starting point: ",
-            &starting_point_label,
-        ));
-    }
-    snapshot_lines.extend(build_onboard_review_digest_display_lines(config));
-    let snapshot_title = if import_source.is_some() {
-        "detected starting point snapshot"
-    } else {
-        "current setup snapshot"
-    };
-
-    let choices = if include_choices {
-        tui_choices_from_screen_options(&build_onboard_shortcut_screen_options(shortcut_kind))
-    } else {
-        Vec::new()
-    };
-    let default_choice_footer_line = render_shortcut_default_choice_footer_line(shortcut_kind);
-    let footer_lines = append_escape_cancel_hint(vec![default_choice_footer_line]);
-
-    TuiScreenSpec {
-        header_style: TuiHeaderStyle::Compact,
-        subtitle: Some(shortcut_kind.subtitle().to_owned()),
-        title: Some(shortcut_kind.title().to_owned()),
-        progress_line: None,
-        intro_lines: Vec::new(),
-        sections: vec![
-            TuiSectionSpec::Narrative {
-                title: Some(snapshot_title.to_owned()),
-                lines: snapshot_lines,
-            },
-            TuiSectionSpec::Callout {
-                tone: TuiCalloutTone::Success,
-                title: Some("fast lane".to_owned()),
-                lines: vec![shortcut_kind.summary_line().to_owned()],
-            },
-        ],
-        choices,
-        footer_lines,
-    }
-}
-
-fn render_preflight_summary_screen_lines_with_style(
-    checks: &[OnboardCheck],
-    width: usize,
-    flow_style: ReviewFlowStyle,
-    color_enabled: bool,
-) -> Vec<String> {
-    let progress_line = flow_style.progress_line();
-
-    render_preflight_summary_screen_lines_with_progress(
-        checks,
-        width,
-        progress_line.as_str(),
-        color_enabled,
-    )
-}
-
-pub fn render_write_confirmation_screen_lines(
-    config_path: &str,
-    warnings_kept: bool,
-    width: usize,
-) -> Vec<String> {
-    render_write_confirmation_screen_lines_with_style(
-        config_path,
-        warnings_kept,
-        width,
-        ReviewFlowStyle::Guided(GuidedPromptPath::NativePromptPack),
-        false,
-    )
-}
-
-pub fn render_current_setup_write_confirmation_screen_lines(
-    config_path: &str,
-    warnings_kept: bool,
-    width: usize,
-) -> Vec<String> {
-    render_write_confirmation_screen_lines_with_style(
-        config_path,
-        warnings_kept,
-        width,
-        ReviewFlowStyle::QuickCurrentSetup,
-        false,
-    )
-}
-
-pub fn render_detected_setup_write_confirmation_screen_lines(
-    config_path: &str,
-    warnings_kept: bool,
-    width: usize,
-) -> Vec<String> {
-    render_write_confirmation_screen_lines_with_style(
-        config_path,
-        warnings_kept,
-        width,
-        ReviewFlowStyle::QuickDetectedSetup,
-        false,
-    )
-}
-
-fn render_write_confirmation_screen_lines_with_style(
-    config_path: &str,
-    warnings_kept: bool,
-    width: usize,
-    flow_style: ReviewFlowStyle,
-    color_enabled: bool,
-) -> Vec<String> {
-    let spec = build_write_confirmation_screen_spec(config_path, warnings_kept, flow_style);
-
-    render_onboard_screen_spec(&spec, width, color_enabled)
-}
-
-fn build_onboard_choice_screen_spec(
-    header_style: OnboardHeaderStyle,
-    subtitle: &str,
-    title: &str,
-    step: Option<(GuidedOnboardStep, GuidedPromptPath)>,
-    intro_lines: Vec<String>,
-    options: Vec<OnboardScreenOption>,
-    footer_lines: Vec<String>,
-    show_escape_cancel_hint: bool,
-) -> TuiScreenSpec {
-    let resolved_subtitle = screen_subtitle(subtitle);
-    let resolved_progress_line =
-        step.map(|(step, guided_prompt_path)| step.progress_line(guided_prompt_path));
-    let resolved_footer_lines = if show_escape_cancel_hint {
-        append_escape_cancel_hint(footer_lines)
-    } else {
-        footer_lines
-    };
-    let resolved_choices = tui_choices_from_screen_options(&options);
-
-    TuiScreenSpec {
-        header_style: tui_header_style(header_style),
-        subtitle: resolved_subtitle,
-        title: Some(title.to_owned()),
-        progress_line: resolved_progress_line,
-        intro_lines,
-        sections: Vec::new(),
-        choices: resolved_choices,
-        footer_lines: resolved_footer_lines,
-    }
-}
-
-fn build_onboard_input_screen_spec(
-    title: &str,
-    step: GuidedOnboardStep,
-    guided_prompt_path: GuidedPromptPath,
-    context_lines: Vec<String>,
-    hint_lines: Vec<String>,
-) -> TuiScreenSpec {
-    let resolved_footer_lines = append_escape_cancel_hint(hint_lines);
-    let progress_line = step.progress_line(guided_prompt_path);
-
-    TuiScreenSpec {
-        header_style: TuiHeaderStyle::Compact,
-        subtitle: None,
-        title: Some(title.to_owned()),
-        progress_line: Some(progress_line),
-        intro_lines: context_lines,
-        sections: Vec::new(),
-        choices: Vec::new(),
-        footer_lines: resolved_footer_lines,
-    }
-}
-
-fn render_workspace_step_screen_lines_with_style(
-    values: &onboard_workspace::WorkspaceStepValues,
-    width: usize,
-    color_enabled: bool,
-) -> Vec<String> {
-    let spec = TuiScreenSpec {
-        header_style: TuiHeaderStyle::Compact,
-        subtitle: Some(presentation::workspace_step_subtitle().to_owned()),
-        title: Some(presentation::workspace_step_title().to_owned()),
-        progress_line: Some(guided_step_progress_line(OnboardWizardStep::Workspace)),
-        intro_lines: vec![presentation::workspace_step_summary_line().to_owned()],
-        sections: vec![TuiSectionSpec::Narrative {
-            title: Some("workspace paths".to_owned()),
-            lines: vec![
-                onboard_review_value_line(
-                    "sqlite memory path",
-                    &values.sqlite_path.display().to_string(),
-                    values.sqlite_origin,
-                ),
-                onboard_review_value_line(
-                    "tool file root",
-                    &values.file_root.display().to_string(),
-                    values.file_root_origin,
-                ),
-            ],
-        }],
-        choices: Vec::new(),
-        footer_lines: append_escape_cancel_hint(vec![
-            "- press Enter to keep the displayed workspace path".to_owned(),
-            "- type 'back' to return to runtime defaults".to_owned(),
-        ]),
-    };
-
-    render_onboard_screen_spec(&spec, width, color_enabled)
-}
-
-fn render_protocol_step_screen_lines_with_style(
-    values: &onboard_protocols::ProtocolStepValues,
-    width: usize,
-    color_enabled: bool,
-) -> Vec<String> {
-    let mut lines = vec![onboard_display_line(
-        "- ACP: ",
-        if values.acp_enabled {
-            "enabled"
-        } else {
-            "disabled"
-        },
-    )];
-
-    if let Some(acp_backend) = values.acp_backend.as_deref() {
-        if values.acp_enabled {
-            lines.push(onboard_display_line(
-                "- selected ACP backend: ",
-                acp_backend,
-            ));
-        }
-    } else if values.acp_enabled {
-        lines.push("- selected ACP backend: not configured".to_owned());
-    }
-
-    if let Some(summary) = onboard_protocols::bootstrap_mcp_server_summary(
-        values.acp_enabled,
-        &values.bootstrap_mcp_servers,
-    ) {
-        lines.push(onboard_display_line("- bootstrap MCP servers: ", &summary));
-    }
-
-    let spec = TuiScreenSpec {
-        header_style: TuiHeaderStyle::Compact,
-        subtitle: Some(presentation::protocol_step_subtitle().to_owned()),
-        title: Some(presentation::protocol_step_title().to_owned()),
-        progress_line: Some(guided_step_progress_line(OnboardWizardStep::Protocols)),
-        intro_lines: vec![presentation::protocol_step_summary_line().to_owned()],
-        sections: vec![TuiSectionSpec::Narrative {
-            title: Some("active protocol state".to_owned()),
-            lines,
-        }],
-        choices: Vec::new(),
-        footer_lines: append_escape_cancel_hint(vec![
-            render_default_choice_footer_line(
-                "Enter",
-                if values.acp_enabled {
-                    "keep ACP enabled"
-                } else {
-                    "keep ACP disabled"
-                },
-            ),
-            "- type 'back' to return to workspace".to_owned(),
-        ]),
-    };
-
-    render_onboard_screen_spec(&spec, width, color_enabled)
-}
-
-fn build_write_confirmation_screen_spec(
-    config_path: &str,
-    warnings_kept: bool,
-    flow_style: ReviewFlowStyle,
-) -> TuiScreenSpec {
-    let mut intro_lines = Vec::new();
-    let config_line = format!("- config: {config_path}");
-    let status_line = presentation::write_confirmation_status_line(warnings_kept).to_owned();
-
-    intro_lines.push(config_line);
-    intro_lines.push(status_line);
-
-    let choices = vec![
-        TuiChoiceSpec {
-            key: "y".to_owned(),
-            label: presentation::write_confirmation_label().to_owned(),
-            detail_lines: vec![presentation::write_confirmation_detail().to_owned()],
-            recommended: false,
-        },
-        TuiChoiceSpec {
-            key: "n".to_owned(),
-            label: presentation::write_confirmation_cancel_label().to_owned(),
-            detail_lines: vec![presentation::write_confirmation_cancel_detail().to_owned()],
-            recommended: false,
-        },
-    ];
-
-    let default_choice_line = render_default_choice_footer_line(
-        "y",
-        presentation::write_confirmation_default_choice_description(),
-    );
-    let footer_lines = append_escape_cancel_hint(vec![default_choice_line]);
-
-    TuiScreenSpec {
-        header_style: TuiHeaderStyle::Compact,
-        subtitle: None,
-        title: Some(presentation::write_confirmation_title().to_owned()),
-        progress_line: Some(match flow_style {
-            ReviewFlowStyle::Guided(_) => guided_step_progress_line(OnboardWizardStep::Ready),
-            ReviewFlowStyle::QuickCurrentSetup | ReviewFlowStyle::QuickDetectedSetup => {
-                flow_style.progress_line()
-            }
-        }),
-        intro_lines,
-        sections: Vec::new(),
-        choices,
-        footer_lines,
-    }
-}
-
-fn tui_header_style(style: OnboardHeaderStyle) -> TuiHeaderStyle {
-    match style {
-        OnboardHeaderStyle::Compact => TuiHeaderStyle::Compact,
-    }
-}
-
-fn screen_subtitle(subtitle: &str) -> Option<String> {
-    let trimmed_subtitle = subtitle.trim();
-
-    if trimmed_subtitle.is_empty() {
-        return None;
-    }
-
-    Some(trimmed_subtitle.to_owned())
-}
-
-fn push_starting_point_fit_hint(
-    hints: &mut Vec<StartingPointFitHint>,
-    seen: &mut std::collections::BTreeSet<&'static str>,
-    key: &'static str,
-    detail: impl Into<String>,
-    domain: Option<crate::migration::SetupDomainKind>,
-) {
-    if seen.insert(key) {
-        hints.push(StartingPointFitHint {
-            key,
-            detail: detail.into(),
-            domain,
-        });
-    }
-}
-
-fn summarize_direct_starting_point_source_reason(
-    candidate: &ImportCandidate,
-) -> Option<&'static str> {
-    candidate.source_kind.direct_starting_point_reason()
-}
-
-fn collect_starting_point_fit_hints(candidate: &ImportCandidate) -> Vec<StartingPointFitHint> {
-    let mut hints = Vec::new();
-    let mut seen = std::collections::BTreeSet::new();
-
-    if let Some(reason) = summarize_direct_starting_point_source_reason(candidate) {
-        push_starting_point_fit_hint(&mut hints, &mut seen, "direct_source", reason, None);
-    } else if let Some(provider_domain) = candidate
-        .domains
-        .iter()
-        .find(|domain| domain.kind == crate::migration::SetupDomainKind::Provider)
-        && let Some(decision) = provider_domain.decision
-        && let Some(reason) = provider_domain.kind.starting_point_reason(decision)
-    {
-        let key = match decision {
-            crate::migration::types::PreviewDecision::KeepCurrent => "provider_keep",
-            crate::migration::types::PreviewDecision::UseDetected => "provider_detected",
-            crate::migration::types::PreviewDecision::Supplement
-            | crate::migration::types::PreviewDecision::ReviewConflict
-            | crate::migration::types::PreviewDecision::AdjustedInSession => "provider",
-        };
-        push_starting_point_fit_hint(
-            &mut hints,
-            &mut seen,
-            key,
-            reason,
-            Some(crate::migration::SetupDomainKind::Provider),
-        );
-    }
-
-    if let Some(channels_domain) = candidate
-        .domains
-        .iter()
-        .find(|domain| domain.kind == crate::migration::SetupDomainKind::Channels)
-        && let Some(decision) = channels_domain.decision
-        && let Some(reason) = channels_domain.kind.starting_point_reason(decision)
-    {
-        let key = match decision {
-            crate::migration::types::PreviewDecision::Supplement => "channels_add",
-            crate::migration::types::PreviewDecision::UseDetected => "channels_detected",
-            crate::migration::types::PreviewDecision::KeepCurrent
-            | crate::migration::types::PreviewDecision::ReviewConflict
-            | crate::migration::types::PreviewDecision::AdjustedInSession => "channels",
-        };
-        push_starting_point_fit_hint(
-            &mut hints,
-            &mut seen,
-            key,
-            reason,
-            Some(crate::migration::SetupDomainKind::Channels),
-        );
-    } else if !candidate.channel_candidates.is_empty()
-        && let Some(reason) = crate::migration::SetupDomainKind::Channels
-            .starting_point_reason(crate::migration::types::PreviewDecision::Supplement)
-    {
-        push_starting_point_fit_hint(
-            &mut hints,
-            &mut seen,
-            "channels_add",
-            reason,
-            Some(crate::migration::SetupDomainKind::Channels),
-        );
-    }
-
-    if (!candidate.workspace_guidance.is_empty()
-        || candidate.domains.iter().any(|domain| {
-            domain.kind == crate::migration::SetupDomainKind::WorkspaceGuidance
-                && matches!(
-                    domain.decision,
-                    Some(crate::migration::types::PreviewDecision::UseDetected)
-                        | Some(crate::migration::types::PreviewDecision::Supplement)
-                )
-        }))
-        && let Some(reason) = crate::migration::SetupDomainKind::WorkspaceGuidance
-            .starting_point_reason(crate::migration::types::PreviewDecision::UseDetected)
-    {
-        push_starting_point_fit_hint(
-            &mut hints,
-            &mut seen,
-            "workspace_guidance",
-            reason,
-            Some(crate::migration::SetupDomainKind::WorkspaceGuidance),
-        );
-    }
-
-    for (kind, key) in [
-        (crate::migration::SetupDomainKind::Cli, "cli"),
-        (crate::migration::SetupDomainKind::Memory, "memory"),
-        (crate::migration::SetupDomainKind::Tools, "tools"),
-    ] {
-        if hints.len() >= 3 {
-            break;
-        }
-        if candidate.domains.iter().any(|domain| {
-            domain.kind == kind
-                && matches!(
-                    domain.decision,
-                    Some(crate::migration::types::PreviewDecision::UseDetected)
-                        | Some(crate::migration::types::PreviewDecision::Supplement)
-                )
-        }) && let Some(reason) =
-            kind.starting_point_reason(crate::migration::types::PreviewDecision::UseDetected)
-        {
-            push_starting_point_fit_hint(&mut hints, &mut seen, key, reason, Some(kind));
-        }
-    }
-
-    if hints.is_empty() {
-        let source_count = crate::migration::render::candidate_source_rollup_labels(
-            &migration_candidate_from_onboard(candidate),
-        )
-        .len();
-        if source_count > 1 {
-            push_starting_point_fit_hint(
-                &mut hints,
-                &mut seen,
-                "combined_sources",
-                format!("combine {source_count} reusable sources"),
-                None,
-            );
-        }
-    }
-
-    hints
-}
-
-fn format_starting_point_reason(hints: &[StartingPointFitHint]) -> Option<String> {
-    if hints.is_empty() {
-        return None;
-    }
-
-    Some(format!(
-        "good fit: {}",
-        hints
-            .iter()
-            .take(3)
-            .map(|hint| hint.detail.as_str())
-            .collect::<Vec<_>>()
-            .join(" + ")
-    ))
-}
-
-fn should_include_starting_point_domain_decision(candidate: &ImportCandidate) -> bool {
-    candidate.source_kind == crate::migration::ImportSourceKind::RecommendedPlan
-}
-
-fn format_starting_point_domain_detail(
-    candidate: &ImportCandidate,
-    domain: &crate::migration::DomainPreview,
-) -> String {
-    let mut detail = format!("{}: ", domain.kind.label());
-    if should_include_starting_point_domain_decision(candidate)
-        && let Some(decision) = domain.decision
-    {
-        detail.push_str(decision.label());
-        detail.push_str(" · ");
-    }
-    detail.push_str(&domain.summary);
-    detail
-}
-
-fn summarize_starting_point_detail_lines(candidate: &ImportCandidate, width: usize) -> Vec<String> {
-    let mut details = Vec::new();
-    let max_lines = if width < 68 { 4 } else { 5 };
-    let mut detail_lines_used = 0usize;
-    let has_channel_details = !candidate.channel_candidates.is_empty();
-    let has_workspace_guidance_details = !candidate.workspace_guidance.is_empty();
-    let migration_candidate = migration_candidate_from_onboard(candidate);
-    let fit_hints = collect_starting_point_fit_hints(candidate);
-    let emphasized_domains = if width < 68 {
-        fit_hints
-            .iter()
-            .filter_map(|hint| hint.domain)
-            .collect::<std::collections::BTreeSet<_>>()
-    } else {
-        std::collections::BTreeSet::new()
-    };
-
-    if let Some(reason_line) = format_starting_point_reason(&fit_hints) {
-        details.push(reason_line);
-    }
-
-    let mut source_labels =
-        crate::migration::render::candidate_source_rollup_labels(&migration_candidate);
-    if has_workspace_guidance_details {
-        source_labels.retain(|label| label != "workspace guidance");
-    }
-    let should_render_source_summary =
-        if candidate.source_kind == crate::migration::ImportSourceKind::RecommendedPlan {
-            !source_labels.is_empty()
-        } else {
-            source_labels.len() > 1
-        };
-    if should_render_source_summary {
-        details.push(format!("sources: {}", source_labels.join(" + ")));
-        detail_lines_used += 1;
-    }
-
-    for domain in &candidate.domains {
-        if has_channel_details && domain.kind == crate::migration::SetupDomainKind::Channels {
-            continue;
-        }
-        if has_workspace_guidance_details
-            && domain.kind == crate::migration::SetupDomainKind::WorkspaceGuidance
-        {
-            continue;
-        }
-        if emphasized_domains.contains(&domain.kind) {
-            continue;
-        }
-        details.push(format_starting_point_domain_detail(candidate, domain));
-        detail_lines_used += 1;
-        if detail_lines_used >= max_lines {
-            return details;
-        }
-    }
-
-    for channel in &candidate.channel_candidates {
-        details.push(format!(
-            "{}: {}",
-            channel.label.to_ascii_lowercase(),
-            channel.summary
-        ));
-        detail_lines_used += 1;
-        if detail_lines_used >= max_lines {
-            return details;
-        }
-    }
-
-    if details.len() < max_lines && !candidate.workspace_guidance.is_empty() {
-        let files = candidate
-            .workspace_guidance
-            .iter()
-            .filter_map(|guidance| Path::new(&guidance.path).file_name())
-            .map(|name| name.to_string_lossy().to_string())
-            .collect::<Vec<_>>();
-        if !files.is_empty() {
-            details.push(format!("workspace guidance: {}", files.join(", ")));
-        }
-    }
-
-    if details.is_empty() {
-        details.push("ready to use as a starting point".to_owned());
-    }
-
-    details
-}
-
-fn start_fresh_starting_point_detail_lines() -> Vec<String> {
-    vec![
-        presentation::start_fresh_starting_point_fit_line().to_owned(),
-        presentation::start_fresh_starting_point_detail_line().to_owned(),
-    ]
-}
-
-fn render_starting_point_selection_footer_lines(
-    sorted_candidates: &[ImportCandidate],
-) -> Vec<String> {
-    let Some(first_candidate) = sorted_candidates.first() else {
-        return Vec::new();
-    };
-
-    let first_hint = render_default_choice_footer_line(
-        "1",
-        presentation::starting_point_footer_description(first_candidate.source_kind),
-    );
-
-    vec![first_hint]
-}
-
-pub fn render_starting_point_selection_screen_lines(
-    candidates: &[ImportCandidate],
-    width: usize,
-) -> Vec<String> {
-    render_starting_point_selection_screen_lines_with_style(candidates, width, false)
-}
-
-fn render_starting_point_selection_screen_lines_with_style(
-    candidates: &[ImportCandidate],
-    width: usize,
-    color_enabled: bool,
-) -> Vec<String> {
-    let sorted_candidates = sort_starting_point_candidates(candidates.to_vec());
-    let mut options = sorted_candidates
-        .iter()
-        .enumerate()
-        .map(|(index, candidate)| OnboardScreenOption {
-            key: (index + 1).to_string(),
-            label: onboard_starting_point_label(Some(candidate.source_kind), &candidate.source),
-            detail_lines: summarize_starting_point_detail_lines(candidate, width),
-            recommended: matches!(
-                candidate.source_kind,
-                crate::migration::ImportSourceKind::RecommendedPlan
-            ),
-        })
-        .collect::<Vec<_>>();
-    options.push(OnboardScreenOption {
-        key: "0".to_owned(),
-        label: presentation::start_fresh_option_label().to_owned(),
-        detail_lines: start_fresh_starting_point_detail_lines(),
-        recommended: false,
-    });
-    let footer_lines = render_starting_point_selection_footer_lines(&sorted_candidates);
-
-    render_onboard_choice_screen(
-        OnboardHeaderStyle::Compact,
-        width,
-        presentation::starting_point_selection_subtitle(),
-        presentation::starting_point_selection_title(),
-        None,
-        vec![presentation::starting_point_selection_hint().to_owned()],
-        options,
-        footer_lines,
-        true,
-        color_enabled,
-    )
-}
-
-fn render_starting_point_selection_header_lines_with_style(
-    _candidates: &[ImportCandidate],
-    width: usize,
-    color_enabled: bool,
-) -> Vec<String> {
-    render_onboard_choice_screen(
-        OnboardHeaderStyle::Compact,
-        width,
-        presentation::starting_point_selection_subtitle(),
-        presentation::starting_point_selection_title(),
-        None,
-        vec![presentation::starting_point_selection_hint().to_owned()],
-        Vec::new(),
-        Vec::new(),
-        true,
-        color_enabled,
-    )
-}
-
-pub fn render_provider_selection_screen_lines(
-    plan: &crate::migration::ProviderSelectionPlan,
-    width: usize,
-) -> Vec<String> {
-    render_provider_selection_screen_lines_with_style(
-        plan,
-        GuidedPromptPath::NativePromptPack,
-        width,
-        false,
-    )
-}
-
-fn render_provider_selection_screen_lines_with_style(
-    plan: &crate::migration::ProviderSelectionPlan,
-    guided_prompt_path: GuidedPromptPath,
-    width: usize,
-    color_enabled: bool,
-) -> Vec<String> {
-    let intro = provider_selection_intro_lines(plan);
-    let options = plan
-        .imported_choices
-        .iter()
-        .map(|choice| OnboardScreenOption {
-            key: choice.profile_id.clone(),
-            label: provider_kind_display_name(choice.kind).to_owned(),
-            detail_lines: {
-                let mut detail_lines = vec![
-                    format!("source: {}", choice.source),
-                    format!("summary: {}", choice.summary),
-                ];
-                if let Some(selector_detail) =
-                    crate::migration::provider_selection::selector_detail_line(
-                        plan,
-                        &choice.profile_id,
-                        width,
-                    )
-                {
-                    detail_lines.push(selector_detail);
-                }
-                if let Some(transport_summary) = choice.config.preview_transport_summary() {
-                    detail_lines.push(format!("transport: {transport_summary}"));
-                }
-                detail_lines
-            },
-            recommended: Some(choice.profile_id.as_str()) == plan.default_profile_id.as_deref(),
-        })
-        .collect::<Vec<_>>();
-    render_onboard_choice_screen(
-        OnboardHeaderStyle::Compact,
-        width,
-        "choose the current provider",
-        "choose active provider",
-        Some((GuidedOnboardStep::Provider, guided_prompt_path)),
-        intro,
-        options,
-        with_default_choice_footer(
-            crate::migration::guidance_lines(plan, width),
-            render_provider_selection_default_choice_footer_line(plan),
-        ),
-        true,
-        color_enabled,
-    )
-}
-
-fn render_provider_selection_header_lines(
-    plan: &crate::migration::ProviderSelectionPlan,
-    guided_prompt_path: GuidedPromptPath,
-    width: usize,
-) -> Vec<String> {
-    render_onboard_choice_screen(
-        OnboardHeaderStyle::Compact,
-        width,
-        "choose the current provider",
-        "choose active provider",
-        Some((GuidedOnboardStep::Provider, guided_prompt_path)),
-        provider_selection_intro_lines(plan),
-        vec![],
-        vec![],
-        true,
-        true,
-    )
-}
-
-fn provider_selection_intro_lines(plan: &crate::migration::ProviderSelectionPlan) -> Vec<String> {
-    if plan.imported_choices.is_empty() {
-        vec!["pick the provider that should back this setup".to_owned()]
-    } else if plan.requires_explicit_choice {
-        vec!["other detected settings stay merged".to_owned()]
-    } else {
-        vec!["review the detected provider choices for this setup".to_owned()]
-    }
-}
-
-fn render_provider_selection_default_choice_footer_line(
-    plan: &crate::migration::ProviderSelectionPlan,
-) -> Option<String> {
-    if plan.requires_explicit_choice {
-        return None;
-    }
-    let default_profile_id = plan.default_profile_id.as_deref()?;
-    let default_kind = plan
-        .imported_choices
-        .iter()
-        .find(|choice| choice.profile_id == default_profile_id)
-        .map(|choice| choice.kind)
-        .or(plan.default_kind)?;
-    Some(render_default_choice_footer_line(
-        default_profile_id,
-        &format!("the {} provider", provider_kind_display_name(default_kind)),
-    ))
-}
-
-pub fn render_model_selection_screen_lines(
-    config: &mvp::config::LoongClawConfig,
-    width: usize,
-) -> Vec<String> {
-    render_model_selection_screen_lines_with_style(
-        config,
-        config.provider.model.as_str(),
-        GuidedPromptPath::NativePromptPack,
-        width,
-        false,
-        false,
-    )
-}
-
-pub fn render_model_selection_screen_lines_with_default(
-    config: &mvp::config::LoongClawConfig,
-    prompt_default: &str,
-    width: usize,
-) -> Vec<String> {
-    render_model_selection_screen_lines_with_style(
-        config,
-        prompt_default,
-        GuidedPromptPath::NativePromptPack,
-        width,
-        false,
-        false,
-    )
-}
-
-fn render_model_selection_screen_lines_with_style(
-    config: &mvp::config::LoongClawConfig,
-    prompt_default: &str,
-    guided_prompt_path: GuidedPromptPath,
-    width: usize,
-    color_enabled: bool,
-    catalog_models_available: bool,
-) -> Vec<String> {
-    let selection_context =
-        onboarding_model_policy::onboarding_model_selection_context(&config.provider);
-    let current_model = selection_context.current_model;
-    let recommended_model = selection_context.recommended_model;
-    let preferred_fallback_models = selection_context.preferred_fallback_models;
-    let allows_auto_fallback_hint = selection_context.allows_auto_fallback_hint;
-    let mut context_lines = vec![
-        format!(
-            "- provider: {}",
-            crate::provider_presentation::guided_provider_label(config.provider.kind)
-        ),
-        format!("- current model: {current_model}"),
-    ];
-    if let Some(recommended_model) = recommended_model {
-        context_lines.push(format!("- recommended model: {recommended_model}"));
-    }
-    if !preferred_fallback_models.is_empty() {
-        let preferred_fallback_summary = preferred_fallback_models.join(", ");
-        context_lines.push(format!(
-            "- configured preferred fallback: {preferred_fallback_summary}",
-        ));
-    }
-
-    let mut hint_lines = vec![render_model_selection_default_hint_line(
-        config,
-        prompt_default,
-    )];
-    if catalog_models_available {
-        hint_lines.push(
-            "- use arrow keys to browse or type to filter available provider models".to_owned(),
-        );
-        hint_lines.push(
-            "- choose `enter custom model id` if you want to type an override manually".to_owned(),
-        );
-    } else {
-        hint_lines.push("- type any provider model id to override it".to_owned());
-    }
-    if allows_auto_fallback_hint {
-        let preferred_fallback_summary = preferred_fallback_models.join(", ");
-        hint_lines.push(format!(
-            "- type `auto` to let runtime try configured preferred fallbacks first: {preferred_fallback_summary}",
-        ));
-    }
-
-    render_onboard_input_screen(
-        width,
-        "choose model",
-        GuidedOnboardStep::Model,
-        guided_prompt_path,
-        context_lines,
-        hint_lines,
-        color_enabled,
-    )
-}
-
-pub fn render_api_key_env_selection_screen_lines(
-    config: &mvp::config::LoongClawConfig,
-    default_api_key_env: &str,
-    width: usize,
-) -> Vec<String> {
-    render_api_key_env_selection_screen_lines_with_style(
-        config,
-        default_api_key_env,
-        default_api_key_env,
-        GuidedPromptPath::NativePromptPack,
-        width,
-        false,
-    )
-}
-
-pub fn render_api_key_env_selection_screen_lines_with_default(
-    config: &mvp::config::LoongClawConfig,
-    default_api_key_env: &str,
-    prompt_default: &str,
-    width: usize,
-) -> Vec<String> {
-    render_api_key_env_selection_screen_lines_with_style(
-        config,
-        default_api_key_env,
-        prompt_default,
-        GuidedPromptPath::NativePromptPack,
-        width,
-        false,
-    )
-}
-
-fn render_api_key_env_selection_screen_lines_with_style(
-    config: &mvp::config::LoongClawConfig,
-    default_api_key_env: &str,
-    prompt_default: &str,
-    guided_prompt_path: GuidedPromptPath,
-    width: usize,
-    color_enabled: bool,
-) -> Vec<String> {
-    let mut context_lines = vec![format!(
-        "- provider: {}",
-        crate::provider_presentation::guided_provider_label(config.provider.kind)
-    )];
-    if let Some(current_env) = render_configured_provider_credential_source_value(&config.provider)
-    {
-        context_lines.push(format!("- current source: {current_env}"));
-    }
-    if let Some(suggested_source) =
-        provider_credential_policy::render_provider_credential_source_value(Some(
-            default_api_key_env,
-        ))
-    {
-        context_lines.push(format!("- suggested source: {suggested_source}"));
-    }
-
-    let example_env_name =
-        provider_credential_policy::provider_credential_env_hint(&config.provider)
-            .unwrap_or_else(|| "PROVIDER_API_KEY".to_owned());
-    let mut hint_lines = vec![render_api_key_env_selection_default_hint_line(
-        config,
-        default_api_key_env,
-        prompt_default,
-    )];
-    hint_lines.push("- enter an env var name, not the secret value itself".to_owned());
-    hint_lines.push(format!("- example: {example_env_name}"));
-    if prompt_default.trim().is_empty() {
-        if provider_credential_policy::provider_has_inline_credential(&config.provider) {
-            hint_lines.push("- leave this blank to keep inline credentials".to_owned());
-        }
-    } else if provider_supports_blank_api_key_env(config) {
-        hint_lines.push(render_clear_input_hint_line(
-            "clear the configured credential env",
-        ));
-    }
-
-    render_onboard_input_screen(
-        width,
-        "choose credential source",
-        GuidedOnboardStep::CredentialEnv,
-        guided_prompt_path,
-        context_lines,
-        hint_lines,
-        color_enabled,
-    )
-}
-
-fn render_web_search_credential_selection_screen_lines_with_style(
-    config: &mvp::config::LoongClawConfig,
-    provider: &str,
-    prompt_default: &str,
-    guided_prompt_path: GuidedPromptPath,
-    width: usize,
-    color_enabled: bool,
-) -> Vec<String> {
-    let provider_label = web_search_provider_display_name(provider);
-    let mut context_lines = vec![format!("- provider: {provider_label}")];
-    if let Some(current_value) =
-        configured_web_search_provider_credential_source_value(config, provider)
-    {
-        let label = if current_value == "inline api key" {
-            "- current credential: "
-        } else {
-            "- current source: "
-        };
-        context_lines.extend(mvp::presentation::render_wrapped_text_line(
-            label,
-            &current_value,
-            width,
-        ));
-    }
-    if let Some(suggested_env) = mvp::config::web_search_provider_descriptor(provider)
-        .and_then(|descriptor| descriptor.default_api_key_env)
-        .and_then(|env_name| {
-            provider_credential_policy::render_provider_credential_source_value(Some(env_name))
-        })
-    {
-        context_lines.extend(mvp::presentation::render_wrapped_text_line(
-            "- suggested source: ",
-            &suggested_env,
-            width,
-        ));
-    }
-
-    let mut hint_lines = vec![render_web_search_credential_selection_default_hint_line(
-        config,
-        provider,
-        prompt_default,
-    )];
-    hint_lines.push("- enter an env var name, not the secret value itself".to_owned());
-    let example_env_name = mvp::config::web_search_provider_descriptor(provider)
-        .and_then(|descriptor| {
-            descriptor
-                .default_api_key_env
-                .or_else(|| descriptor.api_key_env_names.first().copied())
-        })
-        .unwrap_or("WEB_SEARCH_API_KEY");
-    hint_lines.push(format!("- example: {example_env_name}"));
-    if prompt_default.trim().is_empty()
-        && web_search_provider_has_inline_credential(config, provider)
-    {
-        hint_lines.push("- leave this blank to keep inline credentials".to_owned());
-    }
-    if configured_web_search_provider_secret(config, provider)
-        .map(str::trim)
-        .is_some_and(|value| !value.is_empty())
-    {
-        hint_lines.push(render_clear_input_hint_line(
-            "clear the configured web search credential",
-        ));
-    }
-
-    render_onboard_input_screen(
-        width,
-        "choose web search credential",
-        GuidedOnboardStep::WebSearchProvider,
-        guided_prompt_path,
-        context_lines,
-        hint_lines,
-        color_enabled,
-    )
-}
-
-pub fn render_system_prompt_selection_screen_lines(
-    config: &mvp::config::LoongClawConfig,
-    width: usize,
-) -> Vec<String> {
-    render_system_prompt_selection_screen_lines_with_style(
-        config,
-        config.cli.system_prompt.as_str(),
-        GuidedPromptPath::InlineOverride,
-        width,
-        false,
-    )
-}
-
-pub fn render_system_prompt_selection_screen_lines_with_default(
-    config: &mvp::config::LoongClawConfig,
-    prompt_default: &str,
-    width: usize,
-) -> Vec<String> {
-    render_system_prompt_selection_screen_lines_with_style(
-        config,
-        prompt_default,
-        GuidedPromptPath::InlineOverride,
-        width,
-        false,
-    )
-}
-
-fn render_system_prompt_selection_screen_lines_with_style(
-    config: &mvp::config::LoongClawConfig,
-    prompt_default: &str,
-    guided_prompt_path: GuidedPromptPath,
-    width: usize,
-    color_enabled: bool,
-) -> Vec<String> {
-    let current_prompt = config.cli.system_prompt.trim();
-    let current_prompt_display = if current_prompt.is_empty() {
-        "built-in default".to_owned()
-    } else {
-        current_prompt.to_owned()
-    };
-
-    render_onboard_input_screen(
-        width,
-        "adjust cli behavior",
-        GuidedOnboardStep::PromptCustomization,
-        guided_prompt_path,
-        vec![format!("- current prompt: {current_prompt_display}")],
-        vec![
-            render_system_prompt_selection_default_hint_line(config, prompt_default),
-            if prompt_default.trim().is_empty() {
-                "- leave this blank to use the built-in behavior".to_owned()
-            } else {
-                render_clear_input_hint_line("use the built-in behavior")
-            },
-            ONBOARD_SINGLE_LINE_INPUT_HINT.to_owned(),
-        ],
-        color_enabled,
-    )
-}
-
-pub fn render_personality_selection_screen_lines(
-    config: &mvp::config::LoongClawConfig,
-    width: usize,
-) -> Vec<String> {
-    render_personality_selection_screen_lines_with_style(
-        config,
-        config.cli.resolved_personality(),
-        width,
-        false,
-    )
-}
-
-fn render_personality_selection_screen_lines_with_style(
-    config: &mvp::config::LoongClawConfig,
-    default_personality: mvp::prompt::PromptPersonality,
-    width: usize,
-    color_enabled: bool,
-) -> Vec<String> {
-    let options = [
-        (
-            mvp::prompt::PromptPersonality::CalmEngineering,
-            "calm engineering",
-            "rigorous, direct, and technically grounded",
-        ),
-        (
-            mvp::prompt::PromptPersonality::FriendlyCollab,
-            "friendly collab",
-            "warm, cooperative, and explanatory when helpful",
-        ),
-        (
-            mvp::prompt::PromptPersonality::AutonomousExecutor,
-            "autonomous executor",
-            "decisive, high-initiative, and execution-oriented",
-        ),
-    ]
-    .into_iter()
-    .map(|(personality, label, detail)| OnboardScreenOption {
-        key: prompt_personality_id(personality).to_owned(),
-        label: label.to_owned(),
-        detail_lines: vec![detail.to_owned()],
-        recommended: personality == default_personality,
-    })
-    .collect::<Vec<_>>();
-
-    render_onboard_choice_screen(
-        OnboardHeaderStyle::Compact,
-        width,
-        "choose how LoongClaw should speak and take initiative",
-        "choose personality",
-        Some((
-            GuidedOnboardStep::Personality,
-            GuidedPromptPath::NativePromptPack,
-        )),
-        vec![format!(
-            "- current personality: {}",
-            prompt_personality_id(config.cli.resolved_personality())
-        )],
-        options,
-        vec![render_default_choice_footer_line(
-            prompt_personality_id(default_personality),
-            "the current personality",
-        )],
-        true,
-        color_enabled,
-    )
-}
-
-fn render_personality_selection_header_lines(
-    config: &mvp::config::LoongClawConfig,
-    width: usize,
-) -> Vec<String> {
-    render_onboard_choice_screen(
-        OnboardHeaderStyle::Compact,
-        width,
-        "choose how LoongClaw should speak and take initiative",
-        "choose personality",
-        Some((
-            GuidedOnboardStep::Personality,
-            GuidedPromptPath::NativePromptPack,
-        )),
-        vec![format!(
-            "- current personality: {}",
-            prompt_personality_id(config.cli.resolved_personality())
-        )],
-        vec![],
-        vec![],
-        true,
-        true,
-    )
-}
-
-pub fn render_prompt_addendum_selection_screen_lines(
-    config: &mvp::config::LoongClawConfig,
-    width: usize,
-) -> Vec<String> {
-    render_prompt_addendum_selection_screen_lines_with_style(config, width, false)
-}
-
-fn render_prompt_addendum_selection_screen_lines_with_style(
-    config: &mvp::config::LoongClawConfig,
-    width: usize,
-    color_enabled: bool,
-) -> Vec<String> {
-    let current_addendum = config
-        .cli
-        .system_prompt_addendum
-        .as_deref()
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .unwrap_or("none");
-
-    render_onboard_input_screen(
-        width,
-        "adjust prompt addendum",
-        GuidedOnboardStep::PromptCustomization,
-        GuidedPromptPath::NativePromptPack,
-        vec![
-            format!(
-                "- personality: {}",
-                prompt_personality_id(config.cli.resolved_personality())
-            ),
-            format!("- current addendum: {current_addendum}"),
-        ],
-        vec![
-            "- press Enter to keep current addendum".to_owned(),
-            "- type '-' to clear it".to_owned(),
-            ONBOARD_SINGLE_LINE_INPUT_HINT.to_owned(),
-        ],
-        color_enabled,
-    )
-}
-
-pub fn render_memory_profile_selection_screen_lines(
-    config: &mvp::config::LoongClawConfig,
-    width: usize,
-) -> Vec<String> {
-    render_memory_profile_selection_screen_lines_with_style(
-        config,
-        config.memory.profile,
-        GuidedPromptPath::NativePromptPack,
-        width,
-        false,
-    )
-}
-
-fn render_memory_profile_selection_screen_lines_with_style(
-    config: &mvp::config::LoongClawConfig,
-    default_profile: mvp::config::MemoryProfile,
-    guided_prompt_path: GuidedPromptPath,
-    width: usize,
-    color_enabled: bool,
-) -> Vec<String> {
-    let options = MEMORY_PROFILE_CHOICES
-        .into_iter()
-        .map(|(profile, label, detail)| OnboardScreenOption {
-            key: memory_profile_id(profile).to_owned(),
-            label: label.to_owned(),
-            detail_lines: vec![detail.to_owned()],
-            recommended: profile == default_profile,
-        })
-        .collect::<Vec<_>>();
-
-    render_onboard_choice_screen(
-        OnboardHeaderStyle::Compact,
-        width,
-        "choose how much memory context LoongClaw should inject",
-        "choose memory profile",
-        Some((GuidedOnboardStep::MemoryProfile, guided_prompt_path)),
-        vec![format!(
-            "- current profile: {}",
-            memory_profile_id(config.memory.profile)
-        )],
-        options,
-        vec![render_default_choice_footer_line(
-            memory_profile_id(default_profile),
-            "the current memory profile",
-        )],
-        true,
-        color_enabled,
-    )
-}
-
-fn render_memory_profile_selection_header_lines(
-    config: &mvp::config::LoongClawConfig,
-    guided_prompt_path: GuidedPromptPath,
-    width: usize,
-) -> Vec<String> {
-    render_onboard_choice_screen(
-        OnboardHeaderStyle::Compact,
-        width,
-        "choose how much memory context LoongClaw should inject",
-        "choose memory profile",
-        Some((GuidedOnboardStep::MemoryProfile, guided_prompt_path)),
-        vec![format!(
-            "- current profile: {}",
-            memory_profile_id(config.memory.profile)
-        )],
-        vec![],
-        vec![],
-        true,
-        true,
-    )
-}
-
-pub fn render_existing_config_write_screen_lines(config_path: &str, width: usize) -> Vec<String> {
-    render_existing_config_write_screen_lines_with_style(config_path, width, false)
-}
-
-fn render_existing_config_write_screen_lines_with_style(
-    config_path: &str,
-    width: usize,
-    color_enabled: bool,
-) -> Vec<String> {
-    render_onboard_choice_screen(
-        OnboardHeaderStyle::Compact,
-        width,
-        "decide how to write the config",
-        "existing config found",
-        None,
-        vec![
-            format!("- config: {config_path}"),
-            "- choose whether to replace it, keep a backup, or cancel".to_owned(),
-        ],
-        build_existing_config_write_screen_options(),
-        vec![render_default_choice_footer_line(
-            "b",
-            "create backup and replace",
-        )],
-        true,
-        color_enabled,
-    )
-}
-
-fn render_existing_config_write_header_lines_with_style(
-    config_path: &str,
-    width: usize,
-    color_enabled: bool,
-) -> Vec<String> {
-    render_onboard_choice_screen(
-        OnboardHeaderStyle::Compact,
-        width,
-        "decide how to write the config",
-        "existing config found",
-        None,
-        vec![
-            format!("- config: {config_path}"),
-            "- choose whether to replace it, keep a backup, or cancel".to_owned(),
-        ],
-        Vec::new(),
-        Vec::new(),
-        true,
-        color_enabled,
-    )
-}
-
-fn onboard_display_line(prefix: &str, value: &str) -> String {
-    format!("{prefix}{value}")
-}
-
-fn review_value_origin_label(origin: OnboardValueOrigin) -> &'static str {
-    match origin {
-        OnboardValueOrigin::CurrentSetup => presentation::current_value_label(),
-        OnboardValueOrigin::DetectedStartingPoint => presentation::detected_value_label(),
-        OnboardValueOrigin::UserSelected => presentation::user_override_label(),
-    }
-}
-
-fn onboard_review_value_line(
-    label: &str,
-    value: &str,
-    origin: Option<OnboardValueOrigin>,
-) -> String {
-    match origin {
-        Some(origin) => format!("- {label} ({}): {value}", review_value_origin_label(origin)),
-        None => format!("- {label}: {value}"),
-    }
-}
-
-fn draft_output_path_origin(draft: &OnboardDraft) -> Option<OnboardValueOrigin> {
-    if draft.output_path.exists() {
-        return Some(OnboardValueOrigin::CurrentSetup);
-    }
-
-    None
-}
-
-fn build_onboard_review_digest_display_lines_for_draft(draft: &OnboardDraft) -> Vec<String> {
-    let mut lines = vec![
-        onboard_review_value_line(
-            "config output path",
-            &draft.output_path.display().to_string(),
-            draft_output_path_origin(draft),
-        ),
-        onboard_review_value_line(
-            "sqlite memory path",
-            &draft.workspace.sqlite_path.display().to_string(),
-            draft.origin_for(OnboardDraft::WORKSPACE_SQLITE_PATH_KEY),
-        ),
-        onboard_review_value_line(
-            "tool file root",
-            &draft.workspace.file_root.display().to_string(),
-            draft.origin_for(OnboardDraft::WORKSPACE_FILE_ROOT_KEY),
-        ),
-    ];
-    lines.extend(build_onboard_protocol_review_digest_display_lines_for_draft(draft));
-    lines.extend(build_onboard_review_digest_display_lines_without_protocols(
-        &draft.config,
-    ));
-    lines
-}
-
-fn build_onboard_protocol_review_digest_display_lines_for_draft(
-    draft: &OnboardDraft,
-) -> Vec<String> {
-    let protocol_values = onboard_protocols::derive_protocol_step_values(draft);
-    let mut lines = vec![onboard_display_line(
-        "- ACP: ",
-        if protocol_values.acp_enabled {
-            "enabled"
-        } else {
-            "disabled"
-        },
-    )];
-
-    if let Some(acp_backend) = protocol_values.acp_backend.as_deref() {
-        if protocol_values.acp_enabled {
-            lines.push(onboard_review_value_line(
-                "ACP backend",
-                acp_backend,
-                protocol_values.acp_backend_origin,
-            ));
-        }
-    } else if protocol_values.acp_enabled {
-        lines.push(onboard_review_value_line(
-            "ACP backend",
-            "not configured",
-            protocol_values.acp_backend_origin,
-        ));
-    }
-
-    if let Some(summary) = onboard_protocols::bootstrap_mcp_server_summary(
-        protocol_values.acp_enabled,
-        &protocol_values.bootstrap_mcp_servers,
-    ) {
-        lines.push(onboard_review_value_line(
-            "bootstrap MCP servers",
-            &summary,
-            protocol_values.bootstrap_mcp_servers_origin,
-        ));
-    }
-
-    lines
-}
-
-fn build_onboard_review_digest_display_lines(config: &mvp::config::LoongClawConfig) -> Vec<String> {
-    let mut lines = build_onboard_review_digest_display_lines_without_protocols(config);
-    lines.extend(build_onboard_protocol_review_digest_display_lines(config));
-    lines
-}
-
-fn build_onboard_review_digest_display_lines_without_protocols(
-    config: &mvp::config::LoongClawConfig,
-) -> Vec<String> {
-    let mut lines = crate::provider_presentation::provider_profile_state_display_lines(
-        config,
-        Some("- provider: "),
-    );
-    lines.push(onboard_display_line("- model: ", &config.provider.model));
-    lines.push(onboard_display_line(
-        "- transport: ",
-        &config.provider.transport_readiness().summary,
-    ));
-
-    if let Some(provider_endpoint) = config.provider.region_endpoint_note() {
-        lines.push(onboard_display_line(
-            "- provider endpoint: ",
-            &provider_endpoint,
-        ));
-    }
-
-    if let Some(credential_line) = render_onboard_review_credential_line(&config.provider) {
-        lines.push(credential_line);
-    }
-
-    let prompt_mode = summarize_prompt_mode(config);
-    lines.push(onboard_display_line("- prompt mode: ", &prompt_mode));
-
-    if config.cli.uses_native_prompt_pack() {
-        lines.push(onboard_display_line(
-            "- personality: ",
-            prompt_personality_id(config.cli.resolved_personality()),
-        ));
-
-        if let Some(prompt_addendum) = summarize_prompt_addendum(config) {
-            lines.push(onboard_display_line(
-                "- prompt addendum: ",
-                &prompt_addendum,
-            ));
-        }
-    }
-
-    lines.push(onboard_display_line(
-        "- memory profile: ",
-        memory_profile_id(config.memory.profile),
-    ));
-
-    let web_search_provider =
-        web_search_provider_display_name(config.tools.web_search.default_provider.as_str());
-    lines.push(onboard_display_line("- web search: ", &web_search_provider));
-
-    if let Some(web_search_credential) = summarize_web_search_provider_credential(
-        config,
-        config.tools.web_search.default_provider.as_str(),
-    ) {
-        let credential_prefix = format!("- {}: ", web_search_credential.label);
-        lines.push(onboard_display_line(
-            &credential_prefix,
-            &web_search_credential.value,
-        ));
-    }
-
-    let enabled_channels = enabled_channel_ids(config)
-        .into_iter()
-        .filter(|channel| channel != "cli")
-        .collect::<Vec<_>>();
-    if !enabled_channels.is_empty() {
-        lines.push(onboard_display_line(
-            "- channels: ",
-            &enabled_channels.join(", "),
-        ));
-    }
-
-    lines
-}
-
-fn build_onboard_protocol_review_digest_display_lines(
-    config: &mvp::config::LoongClawConfig,
-) -> Vec<String> {
-    let protocols = onboard_protocols::protocol_draft_from_config(config);
-    let mut lines = vec![onboard_display_line(
-        "- ACP: ",
-        if protocols.acp_enabled {
-            "enabled"
-        } else {
-            "disabled"
-        },
-    )];
-
-    if let Some(acp_backend) = protocols.acp_backend.as_deref() {
-        if protocols.acp_enabled {
-            lines.push(onboard_display_line("- ACP backend: ", acp_backend));
-        }
-    } else if protocols.acp_enabled {
-        lines.push("- ACP backend: not configured".to_owned());
-    }
-
-    if let Some(summary) = onboard_protocols::bootstrap_mcp_server_summary(
-        protocols.acp_enabled,
-        &protocols.bootstrap_mcp_servers,
-    ) {
-        lines.push(onboard_display_line("- bootstrap MCP servers: ", &summary));
-    }
-
-    lines
-}
-
-fn render_onboard_review_credential_line(provider: &mvp::config::ProviderConfig) -> Option<String> {
-    summarize_provider_credential(provider)
-        .map(|credential| format!("- {}: {}", credential.label, credential.value))
-}
-
-pub(crate) fn summarize_prompt_mode(config: &mvp::config::LoongClawConfig) -> String {
-    if config.cli.uses_native_prompt_pack() {
-        return "native prompt pack".to_owned();
-    }
-
-    "inline system prompt override".to_owned()
-}
-
-pub(crate) fn summarize_prompt_addendum(config: &mvp::config::LoongClawConfig) -> Option<String> {
-    config
-        .cli
-        .system_prompt_addendum
-        .as_deref()
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(str::to_owned)
-}
-
-pub(crate) fn summarize_provider_credential(
-    provider: &mvp::config::ProviderConfig,
-) -> Option<OnboardingCredentialSummary> {
-    if secret_ref_has_inline_literal(provider.oauth_access_token.as_ref()) {
-        return Some(OnboardingCredentialSummary {
-            label: "credential",
-            value: "inline oauth token".to_owned(),
-        });
-    }
-    if let Some(configured_env) = render_configured_provider_credential_source_value(provider) {
-        return Some(OnboardingCredentialSummary {
-            label: "credential source",
-            value: configured_env,
-        });
-    }
-    if secret_ref_has_inline_literal(provider.api_key.as_ref()) {
-        return Some(OnboardingCredentialSummary {
-            label: "credential",
-            value: "inline api key".to_owned(),
-        });
-    }
-    provider_credential_policy::preferred_provider_credential_env_binding(provider)
-        .and_then(|binding| {
-            provider_credential_policy::render_provider_credential_source_value(Some(
-                binding.env_name.as_str(),
-            ))
-        })
-        .map(|credential_env| OnboardingCredentialSummary {
-            label: "credential source",
-            value: credential_env,
-        })
-}
-
-fn provider_supports_blank_api_key_env(config: &mvp::config::LoongClawConfig) -> bool {
-    provider_credential_policy::provider_has_inline_credential(&config.provider)
-        || provider_credential_policy::provider_has_configured_credential_env(&config.provider)
-}
-
-fn prompt_import_candidate_choice(
-    candidates: &[ImportCandidate],
-    width: usize,
-) -> CliResult<Option<usize>> {
-    let screen_options = build_starting_point_selection_screen_options(candidates, width);
-    let idx = select_screen_option("Starting point", &screen_options, Some("1"))?;
-    let selected = screen_options
-        .get(idx)
-        .ok_or_else(|| format!("starting point selection index {idx} out of range"))?;
-    if selected.key == "0" {
-        return Ok(None);
-    }
-    selected
-        .key
-        .parse::<usize>()
-        .map(|value| Some(value - 1))
-        .map_err(|error| {
-            format!(
-                "invalid starting point selection key {}: {error}",
-                selected.key
-            )
-        })
-}
-
-fn prompt_onboard_shortcut_choice(
-    shortcut_kind: OnboardShortcutKind,
-) -> CliResult<OnboardShortcutChoice> {
-    let options = build_onboard_shortcut_screen_options(shortcut_kind);
-    match select_screen_option("Your choice", &options, Some("1"))? {
-        0 => Ok(OnboardShortcutChoice::UseShortcut),
-        1 => Ok(OnboardShortcutChoice::AdjustSettings),
-        idx => Err(format!("shortcut selection index {idx} out of range")),
-    }
-}
-
-pub fn detect_import_starting_config_with_channel_readiness(
-    readiness: ChannelImportReadiness,
-) -> mvp::config::LoongClawConfig {
-    crate::migration::detect_import_starting_config_with_channel_readiness(to_migration_readiness(
-        readiness,
-    ))
-}
-
-fn resolve_channel_import_readiness(
-    config: &mvp::config::LoongClawConfig,
-) -> ChannelImportReadiness {
-    crate::migration::resolve_channel_import_readiness_from_config(config)
-}
-
-fn default_codex_config_paths() -> Vec<PathBuf> {
-    crate::migration::discovery::default_detected_codex_config_paths()
-}
-
-fn to_migration_readiness(
-    readiness: ChannelImportReadiness,
-) -> crate::migration::ChannelImportReadiness {
-    readiness
-}
-
-fn import_surface_from_migration(surface: crate::migration::ImportSurface) -> ImportSurface {
-    ImportSurface {
-        name: surface.name,
-        domain: surface.domain,
-        level: match surface.level {
-            crate::migration::ImportSurfaceLevel::Ready => ImportSurfaceLevel::Ready,
-            crate::migration::ImportSurfaceLevel::Review => ImportSurfaceLevel::Review,
-            crate::migration::ImportSurfaceLevel::Blocked => ImportSurfaceLevel::Blocked,
-        },
-        detail: surface.detail,
-    }
-}
-
-fn import_surface_to_migration(surface: &ImportSurface) -> crate::migration::ImportSurface {
-    crate::migration::ImportSurface {
-        name: surface.name,
-        domain: surface.domain,
-        level: match surface.level {
-            ImportSurfaceLevel::Ready => crate::migration::ImportSurfaceLevel::Ready,
-            ImportSurfaceLevel::Review => crate::migration::ImportSurfaceLevel::Review,
-            ImportSurfaceLevel::Blocked => crate::migration::ImportSurfaceLevel::Blocked,
-        },
-        detail: surface.detail.clone(),
-    }
-}
-
-fn import_candidate_from_migration(
-    candidate: crate::migration::ImportCandidate,
-) -> ImportCandidate {
-    ImportCandidate {
-        source_kind: candidate.source_kind,
-        source: candidate.source,
-        config: candidate.config,
-        surfaces: candidate
-            .surfaces
-            .into_iter()
-            .map(import_surface_from_migration)
-            .collect(),
-        domains: candidate.domains,
-        channel_candidates: candidate.channel_candidates,
-        workspace_guidance: candidate.workspace_guidance,
-    }
-}
-
-fn migration_candidate_from_onboard(
-    candidate: &ImportCandidate,
-) -> crate::migration::ImportCandidate {
-    crate::migration::ImportCandidate {
-        source_kind: candidate.source_kind,
-        source: candidate.source.clone(),
-        config: candidate.config.clone(),
-        surfaces: candidate
-            .surfaces
-            .iter()
-            .map(import_surface_to_migration)
-            .collect(),
-        domains: candidate.domains.clone(),
-        channel_candidates: candidate.channel_candidates.clone(),
-        workspace_guidance: candidate.workspace_guidance.clone(),
-    }
-}
-
-fn migration_candidate_for_onboard_display(
-    candidate: &ImportCandidate,
-) -> crate::migration::ImportCandidate {
-    let mut migration_candidate = migration_candidate_from_onboard(candidate);
-    migration_candidate.source =
-        onboard_starting_point_label(Some(candidate.source_kind), &candidate.source);
-    migration_candidate
-}
-
-fn onboard_starting_point_label(
-    source_kind: Option<crate::migration::ImportSourceKind>,
-    source: &str,
-) -> String {
-    crate::migration::ImportSourceKind::onboarding_label(source_kind, source)
-}
-
-fn detect_render_width() -> usize {
-    mvp::presentation::detect_render_width()
-}
-
-fn enabled_channel_ids(config: &mvp::config::LoongClawConfig) -> Vec<String> {
-    config.enabled_channel_ids()
-}
-
-pub fn validate_non_interactive_risk_gate(
-    non_interactive: bool,
-    accept_risk: bool,
-) -> CliResult<()> {
-    if non_interactive && !accept_risk {
-        return Err(
-            "non-interactive onboarding requires --accept-risk (explicit acknowledgement)"
-                .to_owned(),
-        );
-    }
-    Ok(())
-}
-
-pub fn should_offer_current_setup_shortcut(
-    options: &OnboardCommandOptions,
-    current_setup_state: crate::migration::CurrentSetupState,
-    entry_choice: OnboardEntryChoice,
-) -> bool {
-    !options.non_interactive
-        && entry_choice == OnboardEntryChoice::ContinueCurrentSetup
-        && current_setup_state == crate::migration::CurrentSetupState::Healthy
-        && !onboard_has_explicit_overrides(options)
-}
-
-pub fn should_offer_detected_setup_shortcut(
-    options: &OnboardCommandOptions,
-    entry_choice: OnboardEntryChoice,
-    provider_selection: &crate::migration::ProviderSelectionPlan,
-) -> bool {
-    !options.non_interactive
-        && entry_choice == OnboardEntryChoice::ImportDetectedSetup
-        && !provider_selection.requires_explicit_choice
-        && !onboard_has_explicit_overrides(options)
-}
-
-fn resolve_onboard_shortcut_kind(
-    options: &OnboardCommandOptions,
-    starting_selection: &StartingConfigSelection,
-) -> Option<OnboardShortcutKind> {
-    if should_offer_current_setup_shortcut(
-        options,
-        starting_selection.current_setup_state,
-        starting_selection.entry_choice,
-    ) {
-        return Some(OnboardShortcutKind::CurrentSetup);
-    }
-    if should_offer_detected_setup_shortcut(
-        options,
-        starting_selection.entry_choice,
-        &starting_selection.provider_selection,
-    ) {
-        return Some(OnboardShortcutKind::DetectedSetup);
-    }
-    None
-}
-
-fn secret_ref_has_inline_literal(secret_ref: Option<&SecretRef>) -> bool {
-    let Some(secret_ref) = secret_ref else {
-        return false;
-    };
-
-    secret_ref.inline_literal_value().is_some()
-}
-
-fn onboard_has_explicit_overrides(options: &OnboardCommandOptions) -> bool {
-    option_has_non_empty_value(options.provider.as_deref())
-        || option_has_non_empty_value(options.model.as_deref())
-        || option_has_non_empty_value(options.api_key_env.as_deref())
-        || option_has_non_empty_value(options.web_search_provider.as_deref())
-        || option_has_non_empty_value(options.web_search_api_key_env.as_deref())
-        || option_has_non_empty_value(options.personality.as_deref())
-        || option_has_non_empty_value(options.memory_profile.as_deref())
-        || option_has_non_empty_value(options.system_prompt.as_deref())
-        || option_has_non_empty_value(env::var("LOONGCLAW_WEB_SEARCH_PROVIDER").ok().as_deref())
-}
-
-fn option_has_non_empty_value(raw: Option<&str>) -> bool {
-    raw.is_some_and(|value| !value.trim().is_empty())
-}
-
-fn load_existing_output_config(output_path: &Path) -> Option<mvp::config::LoongClawConfig> {
-    let path_str = output_path.to_str()?;
-    mvp::config::load(Some(path_str))
-        .ok()
-        .map(|(_, config)| config)
-}
-
-pub fn should_skip_config_write(
-    existing_config: Option<&mvp::config::LoongClawConfig>,
-    draft: &mvp::config::LoongClawConfig,
-) -> bool {
-    existing_config.is_some_and(|existing| {
-        if existing == draft {
-            return true;
-        }
-
-        match (mvp::config::render(existing), mvp::config::render(draft)) {
-            (Ok(existing_rendered), Ok(draft_rendered)) => existing_rendered == draft_rendered,
-            _ => false,
-        }
-    })
-}
-
-pub fn parse_provider_kind(raw: &str) -> Option<mvp::config::ProviderKind> {
-    mvp::config::ProviderKind::parse(raw)
-}
-
-pub fn parse_prompt_personality(raw: &str) -> Option<mvp::prompt::PromptPersonality> {
-    match raw.trim().to_ascii_lowercase().as_str() {
-        "calm_engineering" | "engineering" | "calm" => {
-            Some(mvp::prompt::PromptPersonality::CalmEngineering)
-        }
-        "friendly_collab" | "friendly" | "collab" => {
-            Some(mvp::prompt::PromptPersonality::FriendlyCollab)
-        }
-        "autonomous_executor" | "autonomous" | "executor" => {
-            Some(mvp::prompt::PromptPersonality::AutonomousExecutor)
-        }
-        _ => None,
-    }
-}
-
-pub fn parse_memory_profile(raw: &str) -> Option<mvp::config::MemoryProfile> {
-    match raw.trim().to_ascii_lowercase().as_str() {
-        "window_only" | "window" => Some(mvp::config::MemoryProfile::WindowOnly),
-        "window_plus_summary" | "summary" | "summary_window" => {
-            Some(mvp::config::MemoryProfile::WindowPlusSummary)
-        }
-        "profile_plus_window" | "profile" | "profile_window" => {
-            Some(mvp::config::MemoryProfile::ProfilePlusWindow)
-        }
-        _ => None,
-    }
-}
-
-pub fn provider_default_api_key_env(kind: mvp::config::ProviderKind) -> Option<&'static str> {
-    kind.default_api_key_env()
-}
-
-pub fn provider_kind_id(kind: mvp::config::ProviderKind) -> &'static str {
-    kind.as_str()
-}
-
-pub fn provider_kind_display_name(kind: mvp::config::ProviderKind) -> &'static str {
-    kind.display_name()
-}
-
-pub fn prompt_personality_id(personality: mvp::prompt::PromptPersonality) -> &'static str {
-    match personality {
-        mvp::prompt::PromptPersonality::CalmEngineering => "calm_engineering",
-        mvp::prompt::PromptPersonality::FriendlyCollab => "friendly_collab",
-        mvp::prompt::PromptPersonality::AutonomousExecutor => "autonomous_executor",
-    }
-}
-
-pub fn memory_profile_id(profile: mvp::config::MemoryProfile) -> &'static str {
-    match profile {
-        mvp::config::MemoryProfile::WindowOnly => "window_only",
-        mvp::config::MemoryProfile::WindowPlusSummary => "window_plus_summary",
-        mvp::config::MemoryProfile::ProfilePlusWindow => "profile_plus_window",
-    }
-}
-
-pub fn supported_provider_list() -> String {
-    mvp::config::ProviderKind::all_sorted()
-        .iter()
-        .map(|kind| kind.as_str())
-        .collect::<Vec<_>>()
-        .join(", ")
-}
-
-pub fn supported_personality_list() -> &'static str {
-    "calm_engineering, friendly_collab, autonomous_executor"
-}
-
-pub fn supported_memory_profile_list() -> &'static str {
-    "window_only, window_plus_summary, profile_plus_window"
-}
-
-fn resolve_write_plan(
-    output_path: &Path,
-    options: &OnboardCommandOptions,
-
-    context: &OnboardRuntimeContext,
-) -> CliResult<ConfigWritePlan> {
-    if !output_path.exists() {
-        return Ok(ConfigWritePlan {
-            force: false,
-            backup_path: None,
-        });
-    }
-    if options.force {
-        return Ok(ConfigWritePlan {
-            force: true,
-            backup_path: None,
-        });
-    }
-
-    if options.non_interactive {
-        return Err(format!(
-            "config {} already exists (use --force to overwrite)",
-            output_path.display()
-        ));
-    }
-
-    let existing_path = output_path.display().to_string();
-    print_stdout_lines(render_existing_config_write_header_lines_with_style(
-        &existing_path,
-        context.render_width,
-        true,
-    ))?;
-    let options = build_existing_config_write_screen_options();
-    let selected = options
-        .get(select_screen_option("Your choice", &options, Some("b"))?)
-        .ok_or_else(|| "existing-config write selection out of range".to_owned())?;
-    match selected.key.as_str() {
-        "o" => Ok(ConfigWritePlan {
-            force: true,
-            backup_path: None,
-        }),
-        "b" => Ok(ConfigWritePlan {
-            force: true,
-            backup_path: Some(resolve_backup_path(output_path)?),
-        }),
-        "c" => Err("onboarding cancelled: config file already exists".to_owned()),
-        key => Err(format!(
-            "unexpected existing-config write selection key: {key}"
-        )),
-    }
-}
-
 #[cfg(test)]
 mod tests;
